### 2021-08-03 #458 @cromanpa94
Submission: phruta
Submitting Author Name: Cristian Román-Palacios
Submitting Author Github Handle: @cromanpa94
Repository:  https://github.com/cromanpa94/phruta
Version submitted: 0.1.0
Submission type: Standard
Version accepted: TBD

---

```
Package: phruta
Title: Phylogenetic Reconstruction and Time-dating
Version: 0.1.0
Authors@R: person("Cristian", "Roman Palacios", email = "cromanpa@g.ucla.edu",
                  role = c("aut", "cre"))
Description: The phruta R package is designed to simplify the basic
            phylogenetic pipeline. Specifically, all code is run within
            the same program and data from intermediate steps are saved in
            independent folders. Furthrmore, all code is run within the
            same environment which increases the reproducibility of your analysis.
            phruta retrieves gene sequences, combines newly downloaded and local
            gene sequences, and performs sequence alignments.
biocViews: DECIPHER,
           Biostrings
    rgbif,
    pbapply,
    rentrez,
    taxize,
    ape,
    ips,
    geiger,
    DECIPHER,
    Biostrings
    covr,
    markdown,
    utf8
```

`phruta` simplifies the phylogenetic pipeline in R, increases reproducibility, and helps organizing information used to infer molecular phylogenies. `phruta` retrieves sequences from genbank, curates the resulting files, perform multiple sequence alignment, and tree inference/dating.

Researchers, students, instructors, and analysts interesting in assembling phylogenies (and even time-calibrated phylogenies) for specific groups across the Tree of Life.

[phylotaR](https://github.com/ropensci/phylotaR) retrieves orthologous clusters for particular clades. However, it has limited flexibility in terms of sampling outgroup taxa and (to my knowledge) does not perform any type of processing after retrieving the clusters.
