+++ R/tree.constraint.R
@@ -1,0 +1,73 @@
+#' Tree inference under RAxML
+#'
+#' Performs tree inference under \code{"RAxML"} for aligned fasta sequences in
+#' a given folder (default is \code{"2.Alignments"}).
+#'
+#' @param taxonomy_folder Name of the folder where the 1.Taxonomy file is stored.
+#' @param targetColumns Where to find \code{"RAxML"} or how to run it from the console? (string).
+#' @param Topology A string summarizing the desired topological constraint in newick format.
+#' @param outgroup Optional and only requiered when the topology argument is not "((ingroup), outgroup);".
+#'
+#'
+#' @return None
+#'
+#' @examples
+#' \dontrun{
+#' sq.retrieve(
+#'   clades = c("Felis", "Vulpes", "Phoca"),
+#'   species = "Manis_pentadactyla",
+#'   genes = c("ADORA3")
+#' )
+#' sq.curate(
+#'   filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
+#'   kingdom = "animals", folder = "0.Sequences"
+#' )
+#'
+#'tree.constraint(taxonomy_folder="1.CuratedSequences",
+#'                targetColumns=c("kingdom","phylum","class","order","family","genus","species_names"),
+#'                Topology="((ingroup), outgroup);",
+#'                outgroup="Manis_pentadactyla")
+#'tree.constraint(taxonomy_folder="1.CuratedSequences",
+#'                targetColumns=c("kingdom","phylum","class","order","family","genus","species_names"),
+#'                Topology="((Felis), (Phoca));")
+#' }
+#' @export
+
+tree.constraint <- function(taxonomy_folder="1.CuratedSequences",
+                            targetColumns=c("kingdom","phylum","class","order","family","genus","species_names"),
+                            Topology="((ingroup), outgroup);",
+                            outgroup=NULL){
+
+  taxonomy<-read.csv(paste0(taxonomy_folder, '/1.Taxonomy.csv'))
+
+  if(Topology == "((ingroup), outgroup);"){
+    Topology1<-Topology
+    ingroup<-taxonomy[!taxonomy$species_names==outgroup,]
+    outgroup<-taxonomy[taxonomy$species_names==outgroup,]
+    clades<-list("ingroup"= getListConstraints(ingroup, targetColumns, byClades=F), "outgroup" = getListConstraints(outgroup, targetColumns, byClades=F))
+    for(i in seq_along(clades)){
+    Topology<- gsub(names(clades[i]),clades[[i]]  ,Topology)
+  }
+
+  }else{
+    cstByClade<-getListConstraints(taxonomy, byClades=T)
+    Topology1<-Topology
+    TopologyOriginal<-Topology
+    Topology<-sub(';', '', Topology, fixed = T)
+    Topology<- gsub("[()]", "", Topology)
+    Topology<-sub(' ', '', Topology, fixed = T)
+    clades<-strsplit(Topology, ',')[[1]]
+    for(i in seq_along(clades)){
+      TopologyOriginal<- gsub(clades[i], cstByClade[names(cstByClade) ==clades[i] ]  ,TopologyOriginal)
+    }
+    Topology<-TopologyOriginal
+  }
+
+  unlink("3.2.Phylogeny.constraint", recursive = TRUE)
+  dir.create("3.2.Phylogeny.constraint")
+  write(Topology1, '3.2.Phylogeny.constraint/OriginalConstraints.cst.tre')
+  write(Topology, '3.2.Phylogeny.constraint/phruta.cst.tre')
+  tree<-read.tree('3.2.Phylogeny.constraint/phruta.cst.tre')
+  write.tree(tree, '3.2.Phylogeny.constraint/phruta.ape.cst.tre')
+}
+
