+++ man/tree.raxml.Rd
@@ -10,6 +10,7 @@ tree.raxml(
   raxml_exec = "raxmlHPC",
   Bootstrap = 100,
   outgroup,
+  partitioned = F,
   ...
 )
 }
@@ -25,6 +26,8 @@ this argument can be set to \code{"NULL"} if no specific pattern wants to be ana
 
 \item{outgroup}{A single string of comma-separated tip labels to be used as outgroup in
 \code{"RAxML"} See \code{"RAxML"} documentation for more details (character).}
+
+\item{partitioned}{Whether analyses should be partitioned by gene (Logical).}
 
 \item{...}{Arguments passed to \code{"ips::raxml"}.}
 }
