+++ DESCRIPTION
@@ -26,7 +26,8 @@ Imports:
     geiger,
     methods,
     DECIPHER,
-    Biostrings
+    Biostrings,
+    Rogue
 RoxygenNote: 7.1.1
 Suggests: 
     testthat (>= 3.0.0),

[…]
