+++ .github/workflows/pkgdown.yaml
@@ -3,6 +3,7 @@ on:
     branches:
       - main
       - master
+      - dev
     tags:
       -'*'
 
