+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   phruta_vignette: phruta_vignette.html
-last_built: 2021-08-02T04:16Z
+last_built: 2021-08-08T06:53Z
 urls:
   reference: https://cromanpa94.github.io/phruta//reference
   article: https://cromanpa94.github.io/phruta//articles
