+++ R/sq.partitionfinder1.R
@@ -30,7 +30,6 @@
 #' sq.partitionfinder1(
 #'   folderAlignments = "2.Alignments",
 #'   FilePatterns = "Masked",
-#'   folderPartitionFinder = "2.1.PartitionFinderv1",
 #'   models = "all"
 #' )
 #' }
