+++ NAMESPACE
@@ -6,6 +6,7 @@ export(sq.aln)
 export(sq.curate)
 export(sq.partitionfinderv1)
 export(sq.retrieve)
+export(taxonomy.retrieve)
 export(tree.constraint)
 export(tree.dating)
 export(tree.raxml)

[…]
