+++ pkgdown.yml
@@ -2,8 +2,9 @@ pandoc: 2.7.3
 pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
-  phruta_vignette: phruta_vignette.html
-last_built: 2021-08-08T06:53Z
+  Brief_Intro_phruta: Brief_Intro_phruta.html
+  Phruta_advanced: Phruta_advanced.html
+last_built: 2021-08-08T19:56Z
 urls:
   reference: https://cromanpa94.github.io/phruta//reference
   article: https://cromanpa94.github.io/phruta//articles
