+++ R/sq.aln.R
@@ -27,7 +27,7 @@
 #' sq.retrieve(
 #'   clades = c("Felis", "Vulpes", "Phoca"),
 #'   species = "Manis_pentadactyla",
-#'   genes = c("ADORA3")
+#'   genes = c("ADORA3", "CYTB")
 #' )
 #' sq.curate(
 #'   filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
