+++ man/sq.partitionfinderv1.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sq.partitionfinder1.R
+% Please edit documentation in R/sq.partitionfinderv1.R
 \name{sq.partitionfinderv1}
 \alias{sq.partitionfinderv1}
 \title{Run Partitionfinder v.1}
@@ -26,20 +26,21 @@ None
 }
 \description{
 This function runs partitionfinder v1 within phruta. For now, all analyses are based on genes.
+Please note that you need at least two gene regions to run partitionfinder.
 }
 \examples{
 \dontrun{
 sq.retrieve(
   clades = c("Felis", "Vulpes", "Phoca"),
   species = "Manis_pentadactyla",
-  genes = c("ADORA3")
+  genes = c("ADORA3", "CYTB")
 )
 sq.curate(
   filterTaxonomicCriteria = "Felis|Vulpes|Phoca|Manis",
   kingdom = "animals", folder = "0.Sequences"
 )
 sq.aln(folder = "1.CuratedSequences")
-sq.partitionfinder1(
+sq.partitionfinderv1(
   folderAlignments = "2.Alignments",
   FilePatterns = "Masked",
   models = "all"
