### 2021-08-18 #458 @mpadge
@cromanpa94 Please excuse the delayed response. Most of the ROpenSci team were on vacation, and only now able to debug our new package check bot. Current Editor-in-Chief @noamross will respond asap.
