+++ .github/workflows/render-rmarkdown.yaml
@@ -30,7 +30,7 @@ jobs:
 
       - name: Install packages
         run: |
-          R -e 'install.packages("renv")'
+          R -e 'install.packages("renv", "rmarkdown")'
           R -e 'renv::restore()'
 
       - name: Render Rmarkdown files
