### 2021-08-23 #458 @cromanpa94
Hi @mpadge! No worries! I actually made some updates to the package earlier this month (currently on my `dev` branch). I will account for these initial checks by the end of this week.
