### 2021-08-24 #458 @mpadge
@cromanpa94 To make editor @annakrystalli's job easier, can you run [`pkgcheck::pkgcheck()`](https://github.com/ropensci-review-tools/pkgcheck) yourself once you've made appropriate changes and confirm the `summary` at the top (or result of `summary()` applied to check results) gives all :heavy_check_mark: and no :heavy_multiplication_x:? It should also confirm, "This package may be submitted." Thanks!
