### 2021-08-24 4e0ddf4
Dev branch (#2)
* constraint and rogue taxa

* Styler partitionfinder1

* githubactions updated for dev

* Additional vignette

* Examples updated for two gene regions

* rmarkdown for the rmarkdown rworkflow

Co-authored-by: Cristian Román Palacios <cromanpa@Mac-mini.local>
