+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   Brief_Intro_phruta: Brief_Intro_phruta.html
   Phruta_advanced: Phruta_advanced.html
-last_built: 2021-08-24T15:24Z
+last_built: 2021-08-24T15:57Z
 urls:
   reference: https://cromanpa94.github.io/phruta//reference
   article: https://cromanpa94.github.io/phruta//articles
