+++ DESCRIPTION
@@ -6,6 +6,7 @@ Authors@R: person("Cristian", "Roman Palacios", email = "cromanpa@g.ucla.edu",
                   role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-1696-4886"))
 URL: https://github.com/cromanpa94/phruta, https://cromanpa94.github.io/phruta/
+BugReports: https://github.com/cromanpa94/phruta/issues
 Description: The phruta R package is designed to simplify the basic 
             phylogenetic pipeline. Specifically, all code is run within 
             the same program and data from intermediate steps are saved in 
