+++ tests/testthat/testthat.R
@@ -269,15 +269,15 @@ test_that("Tree constraints ingroup/outgroup", {
 
 ##PartitionFinder
 
-test_that("Tree constraints ingroup/outgroup", {
-  expect_snapshot_output(
-sq.partitionfinderv1(
-  folderAlignments = "2.Alignments",
-  FilePatterns = "Masked",
-  models = "all"
-)
-)
-})
+#test_that("Tree constraints ingroup/outgroup", {
+#  expect_snapshot_output(
+#sq.partitionfinderv1(
+#  folderAlignments = "2.Alignments",
+#  FilePatterns = "Masked",
+#  models = "all"
+#)
+#)
+#})
 
 #Tree rogue
 
