+++ R/taxonomy.retrieve.R
@@ -1,84 +1,110 @@
 #' Curate sequences from genbank
 #'
-#' After downloading sequences from genbank, this function curates sequences based on taxonomic
-#' information. Note that this function provides two summary datasets. First, the accession numbers.
-#' Second, the taxonomic information for each species in the database. The taxonomy strictly follows
-#' the gbif taxonomic backbone. The resulting files are saved to \code{"1.CuratedSequences"}. The
-#' resulting files also have the most recent curated taxonomy following the gbif taxonomic backbone.
+#' After downloading sequences from genbank, this function curates sequences
+#' based on taxonomic information. Note that this function provides two summary
+#' datasets. First, the accession numbers. Second, the taxonomic information
+#' for each species in the database. The taxonomy strictly follows
+#' the gbif taxonomic backbone. The resulting files are saved to
+#'\code{"1.CuratedSequences"}. The resulting files also have the most recent
+#'curated taxonomy following the gbif taxonomic backbone.
 #'
 #' @param species_names A vector of species names
-#' @param database A name of a database with taxonomic information. 'gbif' only
-#'                 works for animals and plants. Databases follows taxize::classification
-#' @param kingdom Optional and only used when database='gbif'. Two possible options: "animals" or "plants."
+#' @param database A name of a database with taxonomic information.
+#'                 'gbif' only works for animals and plants. Databases
+#'                 follows taxize::classification
+#' @param kingdom Optional and only used when database='gbif'. Two possible
+#'                options: "animals" or "plants."
 #'
 #' @import rgbif
 #' @import taxize
 #'
-#' @return data.frame of taxonomic information for the target species (valid in the database)
+#' @return data.frame of taxonomic information for the target species
+#'         (valid in the database)
 #'
 #' @examples
 #' \dontrun{
-#' taxonomy.retrieve(species_names=c("Felis_catus", "PREDICTED:_Vulpes",
-#'  "Phoca_largha", "PREDICTED:_Phoca" ,
-#'  "PREDICTED:_Manis" , "Felis_silvestris" , "Felis_nigripes"),
-#'  database='gbif', kingdom='animals')
+#' taxonomy.retrieve(
+#'   species_names = c(
+#'     "Felis_catus", "PREDICTED:_Vulpes",
+#'     "Phoca_largha", "PREDICTED:_Phoca",
+#'     "PREDICTED:_Manis", "Felis_silvestris", "Felis_nigripes"
+#'   ),
+#'   database = "gbif", kingdom = "animals"
+#' )
 #' }
 #' @export
 
-taxonomy.retrieve<-function(species_names=NULL, database='gbif',
-                            kingdom=NULL){
+taxonomy.retrieve <-
+  function(species_names = NULL,
+           database = "gbif",
+           kingdom = NULL) {
+    ranks <-
+      c("kingdom",
+        "phylum",
+        "class",
+        "order",
+        "family",
+        "genus",
+        "species")
 
-  ranks <- c("kingdom", "phylum", "class", "order", "family", "genus", "species")
-
-  if(database != "gbif" ){
-
-    taxo<-classification(species_names, db = database, rows = 1)
-    invisible(
-    Taxonomy_species<-as.data.frame(do.call(rbind,lapply(seq_along(taxo), function(x){
-      if(!is.na(taxo[[x]])){
-        t(data.frame(taxo[[x]][taxo[[x]][,2] %in% ranks,1]))
-      }else{
-        sma<-matrix(nrow = 1, ncol = length(ranks)-1)
-        cbind(sma, species_names[x])
-      }
-    })))
[…]
