+++ R/sq.curate.R
@@ -1,18 +1,28 @@
 #' Curate sequences from genbank
 #'
-#' After downloading sequences from genbank, this function curates sequences based on taxonomic
-#' information. Note that this function provides two summary datasets. First, the accession numbers.
-#' Second, the taxonomic information for each species in the database. The taxonomy strictly follows
-#' the gbif taxonomic backbone. The resulting files are saved to \code{"1.CuratedSequences"}. The
-#' resulting files also have the most recent curated taxonomy following the gbif (or selected database) taxonomic backbone.
+#' After downloading sequences from genbank, this function
+#' curates sequences based on taxonomic
+#' information. Note that this function provides two summary datasets.
+#' First, the accession numbers.
+#' Second, the taxonomic information for each species in the database.
+#' The taxonomy strictly follows
+#' the gbif taxonomic backbone. The resulting files are saved
+#'  to \code{"1.CuratedSequences"}. The
+#' resulting files also have the most recent curated taxonomy
+#'  following the gbif (or selected database) taxonomic backbone.
 #'
-#' @param filterTaxonomicCriteria A single string of terms (delimited using "|") listing all the
-#'                                strings that could be used to identify the species that should
+#' @param filterTaxonomicCriteria A single string of terms (delimited using "|")
+#'                                listing all the strings that could be used to
+#'                                identify the species that should
 #'                                be in the dataset (character).
-#' @param database A name of a database with taxonomic information. Although 'gbif' is faster,
-#'                 it only has information for animals and plants. Other databases follow taxize::classification.
-#' @param kingdom Optional and only used when database='gbif'. Two possible options: "animals" or "plants."
-#' @param folder The name of the folder where the original sequences are (character).
+#' @param database A name of a database with taxonomic information.
+#'                 Although 'gbif' is faster, it only has information for
+#'                 animals and plants. Other databases follow
+#'                 taxize::classification.
+#' @param kingdom Optional and only used when database='gbif'.
+#'                Two possible options: "animals" or "plants."
+#' @param folder The name of the folder where the original sequences are
+#'               located (character).
 #'
 #' @import stats
 #' @import utils
@@ -37,25 +47,44 @@
 #' }
 #' @export
 
-sq.curate <- function(filterTaxonomicCriteria = NULL, database = "gbif" , kingdom = NULL, folder = "0.Sequences") {
-  if (is.null(filterTaxonomicCriteria)) stop("Please provide filtering pattern in the filterTaxonomicCriteria argument")
-  if (length(filterTaxonomicCriteria) > 1) stop("Use a single string in the filterTaxonomicCriteria argument. \n SOLUTION: Split multiple criteria using |")
-  if (is.null(folder)) stop("Folder where curated sequences are saved must be provided")
+sq.curate <- function(filterTaxonomicCriteria = NULL,
+                      database = "gbif",
+                      kingdom = NULL,
+                      folder = "0.Sequences") {
+  if (is.null(filterTaxonomicCriteria))
[…]
