+++ tests/testthat/testthat.R
@@ -279,13 +279,12 @@ test_that("Error when running tree rogue", {
   )
 })
 
-data <- taxonomy.retrieve(
-  species_names = c("Felis_catus", "PREDICTED:_Vulpes",
-                   "Phoca_largha", "PREDICTED:_Phoca" ,
-                   "PREDICTED:_Manis" , "Felis_silvestris" , "Felis_nigripes"),
-  database = 'itis')
 
 test_that("Taxonomy out of gbif", {
   expect_true(
-    class( data) == "data.frame")
+    class(taxonomy.retrieve(
+      species_names = c("Felis_catus", "PREDICTED:_Vulpes",
+                        "Phoca_largha", "PREDICTED:_Phoca" ,
+                        "PREDICTED:_Manis" , "Felis_silvestris" , "Felis_nigripes"),
+      database = 'itis')) == "data.frame")
 })
