+++ tests/testthat/_snaps/testthat.md
@@ -2,3 +2,11 @@
 
     
 
+# Tree constraints non ingroup/outgroup
+
+    
+
+# Tree constraints ingroup/outgroup
+
+    
+
