+++ tests/testthat/2.1.PartitionFinderv1/partition_finder.cfg
@@ -1,23 +1,0 @@
-## ALIGNMENT FILE ##
-alignment = concatenated.phy;
-
-## BRANCHLENGTHS: linked | unlinked ##
-branchlengths = linked;
-
-## MODELS OF EVOLUTION for PartitionFinder: all | raxml | mrbayes | beast | <list> ##
-##              for PartitionFinderProtein: all_protein | <list> ##
-models = all ;
-
-# MODEL SELECCTION: AIC | AICc | BIC #
-model_selection = BIC;
-
-## DATA BLOCKS: see manual for how to define ##
-[data_blocks]
-GENE_1 = 1 - 354;
-GENE_2 = 355 - 1469;
-
-## SCHEMES, search: all | greedy | rcluster | hcluster | user ##
-[schemes]
-search = greedy;
-
-#user schemes go here if search=user. See manual for how to define.#
