### 2021-09-01 #458 @annakrystalli
### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

### Editor comments

Thanks again for your submission @cromanpa94! We might get confirmation from our review-bot at some point too but in the meantime, I've just finished full local editors checks and there's a few overall minor issues to sort out before proceeding. Please see below:

---
#### Installation issues

Was getting the following error when attempting to install from GitHub

```r
devtools::install_github("cromanpa94/phruta", dependencies = T, build_vignettes = T)
```

```
Error: Failed to install 'phruta' from GitHub:
  System command 'R' failed, exit status: 1, stdout + stderr (last 10 lines):
E> ** inst
E> ** byte-compile and prepare package for lazy loading
E> Error in loadNamespace(i, c(lib.loc, .libPaths()), versionCheck = vI[[i]]) :
E>   there is no package called ‘GenomeInfoDbData’
E> Calls: <Anonymous> ... loadNamespace -> withRestarts -> withOneRestart -> doWithOneRestart
E> Execution halted
E> ERROR: lazy loading failed for package ‘phruta’
E> * removing ‘/private/var/folders/8p/87cqdx2s34vfvcgh04l6z72w0000gn/T/RtmpaDpus1/Rinst13e512d837390/phruta’
E>       -----------------------------------
E> ERROR: package installation failed
```

Had to install `GenomeInfoDbData` manually from source with `BiocManager::install("GenomeInfoDbData", type = "source")` to successfully install `phruta`.

#### Internal functions

I've noticed that you are exporting all functions but suppressing the creation on manual pages by adding `@keywords internal` (e.g. `getListConstraints`). This is a little bit unusual. Are you sure you want every  function in the package exported? If a function shouldn't really be used by users, it's best not to export it. Here's a useful post by @maelle on this [topic](https://blog.r-hub.io/2019/12/12/internal-functions/) to help you decide! :)

#### Good practice issues

Some overall minor issues raised by `goodpractice::gp`

── GP phruta ──────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ avoid long code lines, it is bad for readability. Also, many people
    prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    R/sq.aln.R:9:1
    R/SW.phruta.R:21:1
    R/tree.raxml.R:4:1
    R/tree.raxml.R:7:1
    R/tree.raxml.R:8:1
    ... and 10 more lines

  ✖ avoid calling setwd(), it changes the global environment. If you
    need it, consider using on.exit() to restore the working directory.

    R/tree.raxml.R:70:5
    R/tree.raxml.R:93:9
    R/tree.raxml.R:114:9
    R/tree.raxml.R:118:5

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
    1:NCOL(...) expressions. They are error prone and result 1:0 if the
    expression on the right hand side is zero. Use seq_len() or seq_along()
    instead.

    R/getListConstraints.R:7:28
    R/getListConstraints.R:38:15
    R/getListConstraints.R:48:19
    R/getListConstraints.R:57:19
    R/getListConstraints.R:69:15
    ... and 3 more lines

  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific functions
    you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the
    logicals 'TRUE' and 'FALSE' by default, but are not reserved words and hence
    can be overwritten by the user.  Hence, one should always use 'TRUE' and
    'FALSE' for the logicals.

    R/getListConstraints.R:NA:NA
    R/getListConstraints.R:NA:NA
    R/sq.add.R:NA:NA
    R/sq.add.R:NA:NA
    R/sq.aln.R:NA:NA
    ... and 20 more lines

───────────────────────────────────────────────────────────────────────────────────────────

#### Package coverage

While overall coverage is acceptable, a few R files have very low to no coverage. Is there a reason for this?

```r
> covr::package_coverage(pkg_dir)
R/tree.roguetaxa.R: 0.00%
R/tree.dating.R: 9.38%
R/tree.raxml.R: 12.70%
R/sq.partitionfinderv1.R: 39.39%
R/taxonomy.retrieve.R: 73.44%
R/sq.curate.R: 89.42%
R/sq.add.R: 97.44%
R/sq.retrieve.R: 98.73%
R/getListConstraints.R: 100.00%
R/init.R: 100.00%
R/sq.aln.R: 100.00%
R/tree.constraint.R: 100.00%
```

#### `.DS_Store` files

There are many occasions of `.DS_Store` files in the package source code. These files are harmless but in the end superfluous and it's best practice not to include them in repositories. Could you please remove them and `.DS_Store` to your `.gitignore` files to prevent adding them in future?

---

Let me know if anything is unclear or you need help with anything!
