+++ R/tree.roguetaxa.R
@@ -9,6 +9,7 @@
 #'
 #' @importFrom Rogue RogueTaxa
 #' @importFrom ape read.tree
+#' @importFrom utils write.csv
 #'
 #' @return None
 #'
