+++ R/tree.constraint.R
@@ -12,6 +12,8 @@
 #' @param outgroup Optional and only required when the topology
 #'                 argument is not "((ingroup), outgroup);".
 #'
+#' @importFrom ape read.tree write.tree
+#' @importFrom utils read.csv
 #'
 #' @return None
 #'
@@ -56,19 +58,21 @@ tree.constraint <- function(taxonomy_folder = "1.CuratedSequences",
     ingroup <- taxonomy[!taxonomy$species_names == outgroup, ]
     outgroup <- taxonomy[taxonomy$species_names == outgroup, ]
     clades <- list("ingroup" = getListConstraints(ingroup, targetColumns,
-                                                  byClades = F), "outgroup" =
-                     getListConstraints(outgroup, targetColumns, byClades = F))
+                                                  byClades = FALSE),
+                   "outgroup" =
+                     getListConstraints(outgroup,
+                                        targetColumns, byClades = FALSE))
     for (i in seq_along(clades)) {
       Topology <- gsub(names(clades[i]), clades[[i]], Topology)
     }
   } else {
     cstByClade <- invisible(getListConstraints(taxonomy, targetColumns,
-                                               byClades = T))
+                                               byClades = TRUE))
     Topology1 <- Topology
     TopologyOriginal <- Topology
-    Topology <- sub(";", "", Topology, fixed = T)
+    Topology <- sub(";", "", Topology, fixed = TRUE)
     Topology <- gsub("[()]", "", Topology)
-    Topology <- sub(" ", "", Topology, fixed = T)
+    Topology <- sub(" ", "", Topology, fixed = TRUE)
     clades <- strsplit(Topology, ",")[[1]]
     for (i in seq_along(clades)) {
       TopologyOriginal <- gsub(clades[i],

[…]
