### 2021-09-26 #458 @cromanpa94
Sorry for the delay @annakrystalli! The latest version is in my repo now! Please let me know if you find any other issues before sending it out to review!

**Installation issues**
`GenomeInfoDbData` should now be installed with `phruta`. Thank you so much for pointing this out.

**Internal functions**

Thanks for the advice! I decided not to export the internal functions!

**Good practice issues**

- [x] avoid long code lines: **Fixed (I hope!)!**
- [x] avoid calling setwd(): **I cannot avoid changing the directory for the relevant functions. However, if the analyses run, the function returns to the original directory.**
- [x] avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions: **Fixed (I think!)**
- [x] not import packages as a whole: **I'm now following this suggestion.**
- [x] avoid 'T' and 'F': **Perf!**

**Package coverage** _While overall coverage is acceptable, a few R files have very low to no coverage. Is there a reason for this?_

Correct. Functions with low/intermediate coverage are wrappers to external software (e.g. [raxml](https://cme.h-its.org/exelixis/web/software/raxml/)). I'm doing my best to have 100% coverage in all of the other functions that manipulate data within R.

**.DS_Store files**
.DS_Store removed!
