### 2021-09-28 #458 @annakrystalli
Hello again @cromanpa94 ,

Good to see most issues resolved. There's a few I can still see though:
- `.DS_Store` still in the `R` directory.
- Just noticed that `.Rhistory` and `.Rproj.user` are being version controlled. These will need removing from git also. As you may want to keep the files locally (just don't want them tracked by git), here's a [useful post on how to do that](https://stackoverflow.com/questions/1143796/remove-a-file-from-a-git-repository-without-deleting-it-from-the-local-filesyste). It's best to add `.Rhistory` and `.Rproj.user/*` to your `.gitignore` afterwards.

### notes for reviewers

One note on `setwd`. I appreciate that in some cases it might be necessary and that restoring it with `on.exit()` is a good way.  But I do wonder whether it is necessary and couldn't instead be handled by having an explicit parent directory argument set to default to `"."` (equivalent to the current working directory) and better use of the `file` argument in `ips::raxml()`. The important thing here is that it is clear to the user where files will be written out and ideally give them the option to override the default.

Another thing to think about is the deletion prior to re-writing of complete folders. You will likely have  trouble submitting to CRAN with the current approach because it deletes existing files without asking the user for confirmation. This [goes against CRAN policy](https://cran.r-project.org/submit.html), in particular:
> Packages should not write in the user’s home filespace (including clipboards), nor anywhere else on the file system apart from the R session’s temporary directory (or during installation in the location pointed to by TMPDIR: and such usage should be cleaned up). Installing into the system’s R installation (e.g., scripts to its bin directory) is not allowed.
> Limited exceptions may be allowed in interactive sessions if the package obtains confirmation from the user.

You might want to use/explore how `usethis::write_over()` function handles this and include an `overwrite` argument that y default is set to `FALSE` but can be set to overwrite folders explicitly by the user.

In any case, I won't hold up the review process for these points and will start looking for reviewers. However, it gives you  some time to consider them and I will point these out to the reviewers so we can  get their thoughts too and resolve by the end of  the process. 👍😊
