### 2021-11-19 #458 @raynamharris
None. I have never met the author or contributors.

Partially. The R package itself installed just fine. RAxML installed just fine, thanks to an easy conda install option.

I wasn't able to download and successfully install PATHd-8 or treePL, so instead of running `tree.dating()`, I used `ape`'s `read.tree()` function and plotted the tree with `plot()`

With the exception of `tree.dating`, all the functions I tested worked as expected.

Did not perform any automated tests.

---

The author wrote the `phruta` R package to simplify the basic phylogenetic pipeline. As promised, this R packages brings the functionality of multiple phylogenetic tools into the same environment where they can be used in a pipeline. The files are clearly organized and have standard formats, which makes it easy to use in combination with other tools outside the fruta package.

`sq.retrieve` is my favorite function from this package. I wish I had this 10 years ago when I was searching for sequences in a webbrowser one at a time. I can clearly see how I could incorporate this into a data discovery workflow. I also think would be a useful tool to use when teaching, to allow students to quickly make trees with their genes and species of interest.

The tutorial was easy to follow. However, I found it a little difficult to find detailed information about the functions. It would be nice if some of the descriptions for the functions were listed or described on the home page of the website, so the uses knows more about the capability of the package.
