### 2021-11-26 #458 @cromanpa94
@annakrystalli, that sounds good to me! Thank you so much for your time! I will start working on the next release once all the reviews are in. @raynamharris – your comments are great. I'm so glad you find some functions to be especially useful! I will try to improve the connection between `tree.*` functions. Thank you both (and @fboehm in advance)!
