### 2022-01-07 #458 @fboehm
I've never encountered the authors, and I'm unfamiliar with their previous work.

I think a sentence - or part of a sentence - could be added to clarify the intended audience. The problem that the software addresses is clearly stated.

You might need to move or copy the installation instructions for RAxML, PATHd-8, and treePL to the README.md.

While the vignettes can be rendered successfully with `rmarkdown`, it seems that many or all of the code chunks
are not evaluated when rendering code. Is this intentional? Why? If computing time is a concern, it might be better to read results into the vignette code. For example, add a note to the vignette that the computing time is too long to actually run the code, so you're going to read the saved results - that you could add to the package itself - into the R session. Be sure to continue to include the code that's already in the vignette.

It looks like all exported functions have sufficient Roxygen documentation.

It looks like the examples all are bracketed with code to designate that they not be run. Why is this?

The installation of the R package itself succeeds as documented. However, see my comments above about the installation of RAxML, PATHd-8, and treePL.

I'm not aware of any functional claims. Did I overlook something?

I'm also not aware of performance claims.

What is the purpose of the "tests/testthat/_snaps" directory? It contains a single md file. The "tests/testthat" subdirectory contains seven subdirectories that each contain data files. Is this the best place to store data files for tests? Or would it be more sensible to store the files used in tests with the other data files in this package? I haven't seen other packages that store data in the "testthat" directory.

I see also that `codecov` is at 75%, according to the badge on the Github repository. The collection of tests
feel incomplete. For example, at line 14 of file "tests/testthat/testthat.R", the first argument is "Error in
clades and species". Why is an error expected here? And which error is it? Do you want to specify the expected
error message? At line 57, I don't know what "Max sequence as character" means. Can you elaborate this first
argument?

Also, I might add that the organization of the tests, where all tests are in a single file, made it a little harder to assess the coverage of tests. Would it be better to organize the tests with a single file per exported function?

What happens if a database query returns no results? Is this a problem, and is it detected with your code?

---

`phruta` is a really nice package that accesses several stand-alone software packages
from R for use in phylogenetics analysis.
I encourage the authors to elaborate on the suite of unit tests and to
consider reorganizing the suite of tests from one file into multiple files, perhaps one
file per exported function.

1. `doc` directory - is it needed?

I notice that there are two directories, `doc` and `docs`, with similar names. What is the
motivation for having the two distinct directories with these names? Upon inspecting the
contents, it looks like perhaps `pkgdown` generates the `docs` directory. This leaves me
wondering about the need for the `doc` directory. `doc` seems to contain only vignettes,
despite the presence of a `vignettes` directory. Is it possible to delete the `doc` directory
and its contents?

1. Installing RAxML, PATHd-8, and treePL - is installation platform-specific?

It looks like the authors assume a Mac operating system when writing the installation instructions in the
file `vignettes/Brief_intro_phruta.html`. With much scientific computing occurring on machines with a linux operating system, would it be sensible to specify installation instructions for linux users? Are these software packages also available for Windows?

1. Confusing statement in vignette, `vignettes/Brief_intro_phruta.html`.

The sentence "You may also want to check if RAxML is called using raxmlHPC or something else." is confusing. Can you provide additional context for this statement? Why might RAxML be called something else?

1. `phruta` is misspelled as `pruta` in `vignettes/Brief_intro_phruta.html`.

Under section "Functions in phruta", the text for the third bullet uses `pruta`.

1. Delete remaining .DS_Store files from subdirectories. `R/` contains such a file, for example.

1. I'm less familiar with the use of global variables. Are these necessary and helpful here? Why?
