### 2022-01-07 #458 @fboehm
I hope that my review is helpful for the package authors and that it doesn't come across as overly critical. I expect that there are a variety of ways to organize such a package. Thanks for this opportunity to review. Please let me know if I should elaborate or clarify any of my comments or suggestions.
