### 2022-02-28 a655e42
Basic updates on functions, pre Review
