+++ R/sq.curate.R
@@ -23,12 +23,17 @@
 #'                Two possible options: "animals" or "plants."
 #' @param folder The name of the folder where the original sequences are
 #'               located (character).
+#' @param removeOutliers Whether odseq:odseq should be used to remove outliers
+#' @param minSeqs minimum number of sequences per locus
 #'
 #' @import stats
 #' @import utils
 #' @import ape
 #' @import rgbif
 #' @import taxize
+#' @import odseq
+#' @import msa
+#' @import Biostrings
 #'
 #' @return None
 #'
@@ -50,7 +55,9 @@
 sq.curate <- function(filterTaxonomicCriteria = NULL,
                       database = "gbif",
                       kingdom = NULL,
-                      folder = "0.Sequences") {
+                      folder = "0.Sequences",
+                      removeOutliers = TRUE,
+                      minSeqs=5) {
   if (is.null(filterTaxonomicCriteria))
     stop("Please provide filtering pattern in the
          filterTaxonomicCriteria argument")
@@ -60,7 +67,11 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   if (is.null(folder)) stop("Folder where curated
                             sequences are saved must be provided")
 
-  fastaSeqs <- lapply(list.files(folder, full.names = T), read.FASTA)
+  fastaSeqs <- lapply(list.files(folder, full.names = TRUE), function(x){
+    seqs <- read.FASTA(x)
+    seqs[!duplicated(names(seqs))]
+  })
+
   names(fastaSeqs) <- list.files(folder, full.names = F)
   seqNames <- unlist(lapply(unlist(lapply(fastaSeqs, names)),
                             function(x){
@@ -102,7 +113,7 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
     fastaSeqs <- lapply(fastaSeqs, function(x) {
       x[!names(x) %in% dupDel]
     })
-    AccDat <- AccDat[!duplicated(AccDat[, c(3:4)]), ]
+    AccDat <- AccDat[ !AccDat$OriginalNames %in% dupDel  , ]
   }
 
   Full_dataset <- cbind.data.frame(Taxonomy_species, species_names)
@@ -114,7 +125,7 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
 
   toDel<-AccDat[which(AccDat$Species %nin% Full_dataset$originalSpeciesName), 1]
 
-  # Remove any non-species species
+  # Remove any "non-species" species
   if (length(toDel) > 0) {
     fastaSeqs <- lapply(fastaSeqs, function(x) {
       x[!names(x) %in% toDel]
@@ -122,6 +133,38 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
     AccDat <- AccDat[!AccDat$OriginalNames %in% toDel, ]
   }
 
+
+
+  ##Detect outliers...
+  if(isTRUE(removeOutliers)){
+  cat("\n Removing outliers...\n")
+  fastaSeqsOutDet <- lapply(fastaSeqs, function(x){
+    if(length(x)>minSeqs){
+    seqs <- as.list(as.character(x))
+    seqs <- unlist(lapply(seqs,paste0,collapse=""))
+    seqs <- Biostrings::DNAStringSet(seqs)
+    return(seqs)
+    }
+  })
+
+  fastaSeqsOutDet <- Filter(Negate(is.null), fastaSeqsOutDet)
[…]
