+++ R/tree.raxml.R
@@ -59,7 +59,8 @@ tree.raxml <-
     files_fullNames <-
       list.files(folder, FilePatterns, full.names = T)
     files <- list.files(folder, FilePatterns)
-    seq <- lapply(lapply(files_fullNames, read.FASTA), as.matrix)
+    seq <- lapply(lapply(files_fullNames, read.FASTA), function(x) x[!duplicated(names(x))] )
+    seq <- lapply(seq, as.matrix)
     names(seq) <- files
 
     concatenated <- do.call(cbind.DNAbin, c(seq,
