### 2022-02-28 f355585
Details added to the main functions in the readme
