+++ R/tree.raxml.R
@@ -45,7 +45,7 @@ tree.raxml <-
            raxml_exec = "raxmlHPC",
            Bootstrap = 100,
            outgroup,
-           partitioned = F,
+           partitioned = FALSE,
            ...) {
     if (is.null(folder))
       stop("Please provide folder names")
