### 2022-05-19 91c9459
Adding suggestions to basic new functions
