+++ man/gene.sampling.retrieve.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/acc.gene.sampling.R
+\name{gene.sampling.retrieve}
+\alias{gene.sampling.retrieve}
+\title{Retrieve the distribution of genes for given organism in genbank}
+\usage{
+gene.sampling.retrieve(organism, npar = 2, speciesLevel = TRUE)
+}
+\arguments{
+\item{organism}{The name of a single taxon (character).}
+
+\item{npar}{Number of simultaneous searches (character; optional).}
+
+\item{speciesLevel}{Whether the results should be at the species-level (logical).}
+}
+\value{
+data.frame
+}
+\description{
+Retrieve gene names fora given organism in genbank. This function is
+useful to explore the genes that are widely sampled for a given taxon
+in genbank.
+}
+\examples{
+\dontrun{
+test.spp <- gene.sampling.retrieve(organism = "Puma", speciesLevel=TRUE)
+test.pop <- gene.sampling.retrieve(organism = "Puma", speciesLevel=FALSE)#' }
+}
