+++ man/acc.table.retrieve.Rd
@@ -7,9 +7,11 @@
 acc.table.retrieve(clades, species, genes, speciesLevel)
 }
 \arguments{
-\item{clades}{A vector of clade names (character).}
+\item{clades}{A vector of clade names (character). Note that this can either be
+the name of a clade (e.g. Apis) or the code in NCBI (e.g. txid7459).}
 
-\item{species}{A vector of species names (logical).}
+\item{species}{A vector of species names (logical). Note that this can either be
+the name of a clade (e.g. Apis) or the code in NCBI (e.g. txid7459).}
 
 \item{speciesLevel}{Whether the result should be a species-level dataset (logical).}
 
