+++ DESCRIPTION
@@ -45,7 +45,7 @@ Imports:
     foreach,
     doParallel,
     doSNOW
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.0
 Suggests: 
     testthat (>= 3.0.0),
     covr,

[…]
