### 2022-05-20 c49a163
fixing example and packages for dopar
