+++ NAMESPACE
@@ -15,7 +15,6 @@ export(tree.constraint)
 export(tree.dating)
 export(tree.raxml)
 export(tree.roguetaxa)
-import(Biostrings)
 import(DECIPHER)
 import(ape)
 import(doParallel)
@@ -32,6 +31,7 @@ import(rgbif)
 import(stats)
 import(taxize)
 import(utils)
+importFrom(Biostrings,DNAStringSet)
 importFrom(Biostrings,readDNAStringSet)
 importFrom(Biostrings,writeXStringSet)
 importFrom(Rogue,RogueTaxa)
