### 2022-05-23 abb9795
Retrieving genes basic update
