+++ DESCRIPTION
@@ -34,8 +34,8 @@ Imports:
     pbapply,
     rentrez,
     taxize,
-    ape,
     ips,
+    ape,
     geiger,
     methods,
     DECIPHER,
