### 2022-05-24 ef3ee66
Updates on sq.curate
