### 2022-05-24 5c90412
Updating and partitioning vignette files
