### 2022-05-24 ecb9006
Final updates on the intro vignette
