+++ R/sq.curate.R
@@ -80,15 +80,18 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   ##Merge gene files
   if(!is.null(mergeGeneFiles)){
 
-
     mergedSeqs <- lapply(mergeGeneFiles, function(x){
-      refF <- list.files(folder)
+      refF <- list.files(folder, pattern = '.fasta')
       targetF <- paste0(unlist(x), '.fasta')
       if(all(targetF %in% refF)){
 
       seqs <- lapply(x, function(z) read.FASTA(paste0(folder,"/", z, '.fasta')))
-      unlink(paste0(folder,"/", x, '.fasta')) #remove original files
-      do.call(c,seqs)
+
+      #unlink(paste0(folder,"/", x, '.fasta')) #remove original files
+      file.rename(paste0(folder,"/", x, '.fasta'), paste0(folder,"/", x, '.original.non.merged'))
+
+      comS <- do.call(c,seqs)
+      comS[!duplicated(sub(".*? ", "", names(comS)))]
       }else{
         message("\nFiles ", paste(targetF, collapse = " AND "), ", expected to be merged, not found in ", folder,"\n")
       }
@@ -107,12 +110,14 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
 
   ##
 
-  fastaSeqs <- lapply(list.files(folder, full.names = TRUE), function(x){
+  fastaSeqs <- lapply(list.files(folder, pattern = '.fasta', full.names = TRUE), function(x){
     seqs <- read.FASTA(x)
-    seqs[!duplicated(names(seqs))]
+    seqs <- seqs[!duplicated(names(seqs))]
+    seqs[!duplicated(sub(".*? ", "", names(seqs)))]
   })
 
-  names(fastaSeqs) <- list.files(folder, full.names = F)
+  names(fastaSeqs) <- list.files(folder, pattern = '.fasta', full.names = F)
+
   seqNames <- unlist(lapply(unlist(lapply(fastaSeqs, names)),
                             function(x){
     paste0(strsplit(x, " ")[[1]][c(2:3)], collapse = "_")
@@ -124,7 +129,7 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   AccDat <- data.frame("OriginalNames" = unlist(lapply(fastaSeqs, names)),
                        "AccN" = seqAccN,
                        "Species" = seqNames)
-  AccDat$file <- rep(list.files(folder, full.names = F),
+  AccDat$file <- rep(list.files(folder,pattern = '.fasta', full.names = F),
                      unlist(lapply(fastaSeqs, length)))
 
   species_names <- unique(AccDat$Species)
@@ -137,7 +142,6 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   } else {
     taxonomy.retrieve(species_names = species_names, database = "itis", ranks=ranks)
   }
-
 
   # Remove duplicated species
   if (any(duplicated(AccDat[, c(3:4)]))) {
@@ -221,31 +225,41 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   toRename <-
     Full_dataset[Full_dataset$originalSpeciesName!= Full_dataset$species_names,]
 
+
   ## Export
   unlink("1.CuratedSequences", recursive = TRUE)
   dir.create("1.CuratedSequences")
   invisible(lapply(seq_along(curatedSeqs), function(y) {
     if (length(names(curatedSeqs[[y]])) > minSeqs) {
       ## Original
-      write.FASTA(curatedSeqs[[y]],
+      ta.cu <- curatedSeqs[[y]]
+      ta.cu <- ta.cu[!duplicated(sub(".*? ", "", names(ta.cu)))]
+
[…]
