+++ vignettes/Brief_Intro_phruta.Rmd
@@ -15,6 +15,11 @@ output:
     fig_caption: yes
 ---
 
+```{r include = FALSE}
+knitr::opts_chunk$set(eval = FALSE, warning = FALSE, message = FALSE)
+```
+
+
 # Table of Contents
 
 1.  [What is `phruta`](#intro)
@@ -59,20 +64,18 @@ So far, we have decided the taxonomic make of our analyses. However, we also nee
 
 Now, let's look for the gene regions that are sampled for our target taxa. For this, we will use the `gene.sampling.retrieve` function in `pruta`. The resulting `data.frame`, named `gs.seqs` in this example, will contain the list of full names for genes sampled in genbank for the target taxa.
 
-```{r eval=FALSE}
+```{r }
+start_time <- Sys.time()
+library(phruta)
 gs.seqs <- gene.sampling.retrieve(organism = c("Felis", "Vulpes", "Phoca", "Manis_pentadactyla"), speciesSampling=TRUE)
 head(gs.seqs)
 ```
 
-```{r for testing}
-gs.seqs <- data.frame('a')
-```
-
-For the search terms, `phruta` was able to retrieve `r nrow(gs.seqs)` gene regions . The `gene.sampling.retrieve` provides an estimate of the number of species in genbank (matching the taxonomic criteria of the search term) that have sequences for a given gene region. However, this estimate is only as good as the annotations for genes deposited in genbank. 
+For the search terms, `phruta` was able to retrieve the names for multiple gene regions . The `gene.sampling.retrieve` provides an estimate of the number of species in genbank (matching the taxonomic criteria of the search term) that have sequences for a given gene region. However, this estimate is only as good as the annotations for genes deposited in genbank. 
 
 From here, we will generate a preliminary table summarizing accession numbers for the combination of taxa and gene regions that we're interested in sampling. In fact, not all these accession numbers are expected to be in the final (curated) molecular dataset. Using the `acc.table.retrieve` function, we will assemble a species-level summary of accession numbers (hence the `speciesLevel = TRUE` argument). For simplicity, this tutorial will focus on sampling gene regions that are sampled in >30% of the species (`targetGenes` data.frame).
 
-```{r eval=FALSE}
+```{r }
 targetGenes <- gs.seqs[gs.seqs$PercentOfSampledSpecies>30,]
 
 acc.table <- acc.table.retrieve(
@@ -86,7 +89,7 @@ head(acc.table)
 
 The `acc.table` object is a `data.frame` that is later on going to be used for downloading locally the relevant gene sequences. Feel free to review this dataset, make changes, add new species, samples, etc. The integrity of this dataset is critical for the next steps so please take your time and review it carefully. Let's just make some minor changes to our dataset:
 
-```{r eval=FALSE}
+```{r }
 acc.table$Species <- sub("P.", "Phoca ", acc.table$Species, fixed = T)
 acc.table$Species <- sub("F.", "Felis ", acc.table$Species, fixed = T)
 acc.table$Species <- sub("V.", "Vulpes ", acc.table$Species, fixed = T)
@@ -97,26 +100,27 @@ Now, since we're going to retrieve sequences from genbank, we will use the `sq.r
 
 Now, we still need to retrieve the sequences from the accessions table `acc.table`. The `sq.retrieve.indirect` function will write all the resulting fasta files into the newly created folder `0.Sequences` located in our working directory.
 
-```{r eval=FALSE}
+```{r }
 sq.retrieve.indirect(acc.table)
 ```
 
[…]
