+++ R/acc.retrieve.R
@@ -1,4 +1,4 @@
-#' Retrieve accession numbers and titles for a given search in genbank
+#' Retrieve accession numbers and titles for a genbank search
 #'
 #' Retrieve accession numbers and titles for searches in genbank. This function
 #' is useful for exploring the gene-, population-, and species-level sampling
@@ -23,32 +23,32 @@
 
 acc.retrieve <- function(organism, acc.num = FALSE, gene=NULL, speciesLevel=FALSE, npar=2){
 
-  if( is.null(gene) & speciesLevel ){stop("\nPlease provide the name of a gene region or disable the species-level filtering")}
+  if (is.null(gene) & speciesLevel ) {stop("\nPlease provide the name of a gene region or disable the species-level filtering")}
 
   get_gene = function(x, search, nObs){
     tryCatch({
-
-      recs_summ <- if(nObs==1){
+      recs_summ <- if (nObs == 1) {
        reutils::efetch(search,
-                                   rettype = "docsum",
-                                   retmode = "text")
+                      rettype = "docsum",
+                      retmode = "text")
       }else{
       reutils::efetch(search,
-                                   rettype = "docsum",
-                                   retmode = "text",
-                                   retstart = x, retmax=by)
+                      rettype = "docsum",
+                      retmode = "text",
+                      retstart = x,
+                      retmax = by)
       }
       doc <- reutils::content(recs_summ)
       xml_data <- XML::xmlToList(doc)
       do.call(rbind.data.frame, lapply(xml_data, function(y){
-        cbind.data.frame('Ti'=y[[3]][[1]], "Acc"=y[[2]][[1]])
+        cbind.data.frame('Ti' = y[[3]][[1]], "Acc" = y[[2]][[1]])
       }))
-    }, error=function(e){})
+    }, error = function(e){})
   }
 
 
-  if(!acc.num){
-  base.search <- esearch(term = paste0(organism,"[orgn] ", if(!is.null(gene)){paste0("and " ,gene, "[TITL] ")},
+  if (!acc.num) {
+  base.search <- esearch(term = paste0(organism,"[orgn] ", if (!is.null(gene)) {paste0("and " ,gene, "[TITL] ")},
                                        "NOT sp NOT unverified NOT genome NOT aff NOT cf NOT predicted NOT TSA NOT EST"),
                          db = 'nuccore', usehistory = TRUE)
   }else{
@@ -60,11 +60,11 @@ acc.retrieve <- function(organism, acc.num = FALSE, gene=NULL, speciesLevel=FALS
   xml <- content(base.search, "xml")
   count <- as.numeric(XML::xmlToList(xml)$Count)
 
-  if(count>0){
-    if(!acc.num){
+  if (count > 0) {
+    if (!acc.num) {
     message("\nSequences found for gene ", gene, " and organism ", organism)
       }
-    myCluster <- makeCluster(npar, type="SOCK")
+    myCluster <- makeCluster(npar, type = "SOCK")
     registerDoSNOW(myCluster)
     by = 499
     cuts <- seq(1, count, by)
@@ -76,21 +76,21 @@ acc.retrieve <- function(organism, acc.num = FALSE, gene=NULL, speciesLevel=FALS
                      .packages = "reutils",
                      .options.snow = opts
                      ,.combine = 'rbind'
-    ) %dopar% get_gene(x, search = base.search, nObs=count)
+    ) %dopar% get_gene(x, search = base.search, nObs = count)
 
     Species <- sapply(AccDS[,1], function(z) paste(strsplit(z, " ")[[1]][c(1:2)], collapse = " " ))
[…]
