+++ tests/testthat/testthat.R
@@ -11,10 +11,10 @@ test_that("Class of sample phylogenies", {
   expect_true(class(SW.phruta) == "multiPhylo")
 })
 
-# context("sq.retrieve")
+# context("sq.retrieve.direct")
 
 test_that("Error in clades and species", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = NULL,
     species = NULL,
     genes = NULL,
@@ -24,7 +24,7 @@ test_that("Error in clades and species", {
 })
 
 test_that("Error in default numeric clades", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = 1,
     species = NULL,
     genes = NULL,
@@ -36,7 +36,7 @@ test_that("Error in default numeric clades", {
 
 
 test_that("Error in default numeric species", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = NULL,
     species = 1,
     genes = NULL,
@@ -46,7 +46,7 @@ test_that("Error in default numeric species", {
 })
 
 test_that("No genes", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = "Homo",
     species = "Brassica",
     genes = NULL,
@@ -57,7 +57,7 @@ test_that("No genes", {
 
 
 test_that("Max sequence as character", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = "Homo",
     species = "Brassica",
     genes = "COI",
@@ -68,7 +68,7 @@ test_that("Max sequence as character", {
 
 
 test_that("Max length as character", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = "Homo",
     species = "Brassica",
     genes = "COI",
@@ -78,7 +78,7 @@ test_that("Max length as character", {
 })
 
 test_that("length of maxseqs and maxlength", {
-  expect_error(sq.retrieve(
+  expect_error(sq.retrieve.direct(
     clades = "Homo",
     species = "Brassica",
     genes = "COI",
@@ -88,8 +88,8 @@ test_that("length of maxseqs and maxlength", {
 })
 
 
-test_that("Silent sq.retrieve", {
-  expect_output(sq.retrieve(
+test_that("Silent sq.retrieve.direct", {
+  expect_output(sq.retrieve.direct(
     clades = "Psocus",
     species = NULL,
     genes = "HDFJ",
@@ -139,7 +139,7 @@ test_that("assuming the function runs x2", {
 # context("sq.curate")
 
 test_that("assuming the function runs", {
-  expect_invisible(sq.aln(folder = "1.CuratedSequences", FilePatterns = "renamed", mask = T))
+  expect_invisible(sq.aln(folder = "1.CuratedSequences", FilePatterns = "renamed", mask = TRUE))
 })
 
 
@@ -189,7 +189,7 @@ test_that("sq.add default", {
 unlink(list.dirs("."), recursive = TRUE)
 
 test_that("Retrieve sequences", {
-  expect_output(sq.retrieve(
+  expect_output(sq.retrieve.direct(
     clades = c("Felis", "Vulpes", "Phoca"),
     species = "Manis_pentadactyla",
     genes = c("ADORA3", "CYTB")
@@ -232,7 +232,7 @@ test_that("Generate list of constraints, not by clade", {
   expect_true(
     class(getListConstraints(taxonomy,
       targetColumns = c("kingdom", "phylum", "class", "order", "family", "genus", "species_names"),
-      byClades = F
+      byClades = FALSE
     )) == "list"
   )
 })
[…]
