+++ man/tree.raxml.Rd
@@ -6,7 +6,7 @@
 \usage{
 tree.raxml(
   folder = "2.Alignments",
-  FilePatterns = "Masked",
+  FilePatterns = "Masked_",
   raxml_exec = "raxmlHPC",
   Bootstrap = 100,
   outgroup,
