+++ vignettes/Brief_Intro_phruta.Rmd
@@ -28,15 +28,13 @@ library(knitr)
 
 1.  [What is `phruta`](#intro)
 2.  [Functions in `phruta`](#paragraph0)
-3.  [From taxonomic names to time-calibrated phylogenies](#paragraph1)
-4.  [Installing RAxML](#paragraph2)
-5.  [Installing PATHd-8 and treePL](#paragraph3)
+3.  [From taxonomic names to sequence alignments](#paragraph1)
 
 ## What is `phruta` <a name="intro"></a>
 
 The `phruta` R package is designed to simplify the basic phylogenetic pipeline and allow scientists from different backgrounds to assemble their own phylogenies. All code in `phruta` is run within the same program (`R`) and data from intermediate steps are saved locally in independent folders. Furthermore, all code is run within the same environment which increases the reproducibility of your analysis. `phruta` looks for potentially (phylogenetically) relevant gene regions for a given set of taxa, retrieves gene sequences, could combine newly downloaded and local gene sequences, performs sequence alignments, phylogenetic inference, and tree dating. `phruta` is also able to perform basic phylogenetic inference under `RAxML` on the resulting sequence alignments. The current release allows users to conduct tree dating based on secondary calibrations. `phruta` is essentially a wrapper for alternative `R` packages and software.
 
-## Functions in `phruta`
+## Functions in `phruta` <a name="paragraph0"></a>
 
 The current release of `phruta` includes a set of eight major functions. All eight functions form a pipeline within `phruta` to output a time-calibrated phylogeny. However, users interested in using their own files at any stage can run each function independently.
 
@@ -82,7 +80,7 @@ gs.seqs <- gene.sampling.retrieve(organism = c("Felis", "Vulpes", "Phoca", "Mani
 For the search terms, `phruta` was able to retrieve the names for `r nrow(gs.seqs)` gene regions.
 
 ```{r echo=FALSE}
-kable(gs.seqs, caption = "Table 1. Gene sampling for the search including the genera Felis, Vulpes, Phoca as the ingroup taxa. Manis pentadactyla was used as outgroup.")
+kable(head(gs.seqs), caption = "Gene sampling for the search including the genera Felis, Vulpes, Phoca as the ingroup taxa. Manis pentadactyla was used as outgroup.")
 ```
 
 The `gene.sampling.retrieve` provides an estimate of the number of species in genbank (matching the taxonomic criteria of the search term) that have sequences for a given gene region. However, this estimate is only as good as the annotations for genes deposited in genbank. 
@@ -103,7 +101,8 @@ acc.table <- acc.table.retrieve(
 The `acc.table` object is a `data.frame` that is later on going to be used for downloading locally the relevant gene sequences. In this case, the dataset includes the following information:
 
 ```{r echo=FALSE}
-kable(acc.table, caption = 'Table 2. Accession numbers for the relevant species and gene regions examined in this tutorial.')
+row.names(acc.table) <- NULL
+kable(acc.table, caption = 'Accession numbers for the relevant species and gene regions examined in this tutorial.')
 ```
 
 Feel free to review this dataset, make changes, add new species, samples, etc. The integrity of this dataset is critical for the next steps so please take your time and review it carefully. Let's just make some minor changes to our dataset:
@@ -113,12 +112,13 @@ acc.table$Species <- sub("P.", "Phoca ", acc.table$Species, fixed = TRUE)
 acc.table$Species <- sub("F.", "Felis ", acc.table$Species, fixed = TRUE)
 acc.table$Species <- sub("V.", "Vulpes ", acc.table$Species, fixed = TRUE)
 acc.table$Species <- sub("mitochondrial", "", acc.table$Species)
+row.names(acc.table) <- NULL
 ```
 
 Let's check how the new table looks now...
 
 ```{r echo = FALSE}
[…]
