### 2022-05-28 09d142c
testthat and related updates on vignette etc
