+++ tests/testthat/_snaps/testthat.md
@@ -1,12 +1,0 @@
-# Add sequences
-
-    
-
-# Tree constraints non ingroup/outgroup
-
-    
-
-# Tree constraints ingroup/outgroup
-
-    
-
