### 2022-05-28 e5dc805
Plotting raw vs masked in vignette, aln exclude those where the masking fail, >= for filtering in curate
