+++ R/sq.curate.R
@@ -15,6 +15,10 @@
 #'                                listing all the strings that could be used to
 #'                                identify the species that should
 #'                                be in the dataset (character).
+#' @param mergeGeneFiles A named list, with each element being a character vector
+#'                       indicating the names of the files in \code{"0.Sequences"}
+#'                       that need to be combined into a single fasta file. For instance,
+#'                       you can use this argument to combine CO1 and COI.
 #' @param database A name of a database with taxonomic information.
 #'                 Although 'gbif' is faster, it only has information for
 #'                 animals and plants. Other databases follow
@@ -23,8 +27,12 @@
 #'                Two possible options: "animals" or "plants."
 #' @param folder The name of the folder where the original sequences are
 #'               located (character).
-#' @param removeOutliers Whether odseq:odseq should be used to remove outliers
+#' @param removeOutliers Whether  \code{odseq:odseq} should be used to remove outliers
 #' @param minSeqs minimum number of sequences per locus
+#' @param threshold Relative to \code{odseq::odseq}. Only important if
+#'                  \code{removeOutliers = TRUE}
+#' @param ranks The taxonomic ranks used to examine the taxonomy of the species
+#'              in the \code{0.Sequences} folder.
 #'
 #' @import stats
 #' @import utils
@@ -34,6 +42,8 @@
 #' @import odseq
 #' @import msa
 #' @importFrom Biostrings DNAStringSet
+#'
+#' @name sq.curate
 #'
 #' @return None
 #'
