+++ DESCRIPTION
@@ -45,7 +45,8 @@ Imports:
     doSNOW,
     Biostrings,
     msa,
-    odseq
+    odseq,
+    XML
 RoxygenNote: 7.2.0
 Suggests: 
     testthat (>= 3.0.0),
