+++ NAMESPACE
@@ -16,6 +16,7 @@ export(tree.dating)
 export(tree.raxml)
 export(tree.roguetaxa)
 import(DECIPHER)
+import(XML)
 import(ape)
 import(doParallel)
 import(doSNOW)
