+++ README.md
@@ -34,6 +34,13 @@ library(devtools)
 install_github("cromanpa94/phruta")
 ```
 
+Please make sure that the R packages `msa`, `DECIPHER`, `Biostrings`, and `odseq` are correctly installed. Finally, if you're interested in using the development version of `phruta`, feel free to use the following code:
+
+```
+library(devtools) 
+install_github("cromanpa94/phruta", subdir = "devB")
+```
+
 ## How is `phruta` different?
 
 `phruta` has two core functions. The main applications of these functions is briefly outlined below:
