### 2022-05-30 59bdea9
Adding external software installation details
