+++ README.md
@@ -24,6 +24,16 @@ The main functions in the `phruta` R package allow for a quick mining and curati
 
 `phruta` simplifies the phylogenetic pipeline, increases reproducibility, and helps organizing information used to infer molecular phylogenies.
 
+## How is `phruta` different from other software?
+
+`phruta` has two core functions. The main applications of these functions is briefly outlined below:
+
+- `sq.retrieve*`: This function downloads sequences from genbank (nucleotide database) for particular taxa (taxonomic groups or particular species) and a list of genes. 
+
+- `sq.curate()`: After sequences are downloaded from genbank, this function curates sequences within each of the examined genes by detecting sequence outliers and by using taxonomic information. 
+
+In addition to these two main functions, users will be able to align the downloaded sequences, infer phylogenetic trees, and calibrate phylogenies using additional functions in `phruta`.
+
 
 ## Installing `phruta`
 
@@ -41,15 +51,45 @@ library(devtools)
 install_github("cromanpa94/phruta", subdir = "devB")
 ```
 
-## How is `phruta` different?
+## Installing RAxML <a name="paragraph1"></a>
 
-`phruta` has two core functions. The main applications of these functions is briefly outlined below:
+`RAxML` can be easily installed to the `PATH` using one of the two lines below in `conda`:
 
-- `sq.retrieve`: This function downloads sequences from genbank (nucleotide database) for particular taxa (taxonomic groups or particular species) and a list of genes. 
+```{bash eval=FALSE}
+conda install -c bioconda/label/cf201901 raxml 
+```
 
-- `sq.curate()`: After sequences are downloaded from genbank, this function curates sequences within each of the examined genes by detecting sequence outliers and by using taxonomic information. 
+```{bash eval=FALSE}
+conda install -c bioconda raxml
+```
 
-In addition to these two main functions, users will be able to align the downloaded sequences, infer phylogenetic trees, and calibrate phylogenies using additional functions in `phruta`.
+Open `R` and make sure that the following line doesn't throw an error.
+
+```{r eval=FALSE}
+system("raxmlHPC")
+```
+
+You may also want to check if `RAxML` is called using `raxmlHPC` or something else. This can be set when running `tree.raxml` using the `raxml_exec` argument.
+
+Finally, note that `RStudio` sometimes has issues finding *stuff* in the path while using `system()`.
+
+With issues, if you're running macOS, try running RStudio from the command line. Close RStudio, open the terminal, and type:
+
+```{bash eval=FALSE}
+open /Applications/RStudio.app
+```
+
+## Installing PATHd-8 and treePL <a name="paragraph2"></a>
+
+There are excellent guides for installing `PATHd-8` and `treePL`. Here, I summarize two potentially relevant options.
+
+First, you can use [Brian O'Meara's](https://github.com/bomeara/phydocker/blob/master/Dockerfile) approach for installing `PATHd-8`. I summarized the code [here](https://gist.github.com/cromanpa94/a43bc710a17220f71d796d6590ea7fe4).
+
+Second, you can use homebrew to install `treePL`, thanks to Jonathan Chang.
+
+```{bash eval = F}
+brew install brewsci/bio/treepl
+```
 
 
 ## Running `phruta` from `Rstudio` while using `MacOS`?
