### 2022-05-30 e36a916
Adding details to raxml in the vignette
