### 2022-05-30 #458 @cromanpa94
Sorry for the (extreme) delay @annakrystalli – thought I was going to have some time to work on this during the semester!

I'm done implementing several new functions that improve the functionality of `phruta` based on comments from the reviews and additional feedback I received from collages...I'm right now cleaning up the package and accounting for @fboehm's feedback on the tests.

I'll post my response to everyone's comments within a week! Thank you so much for your patience!
