+++ man/taxonomy.retrieve.Rd
@@ -18,8 +18,10 @@ taxonomy.retrieve(
 'gbif' only works for animals and plants. Databases
 follows taxize::classification}
 
-\item{kingdom}{Optional and only used when database='gbif'. Two possible
+\item{kingdom}{Optional and only used when database = 'gbif'. Two possible
 options: "animals" or "plants."}
+
+\item{ranks}{Taxonomic ranks to retrieve from the target species.}
 }
 \value{
 data.frame of taxonomic information for the target species
