+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   Future_phruta: Future_phruta.html
   Phruta_advanced: Phruta_advanced.html
   Phylogenetics_phruta: Phylogenetics_phruta.html
-last_built: 2022-05-30T04:37Z
+last_built: 2022-05-30T15:14Z
 urls:
   reference: https://cromanpa94.github.io/phruta/reference
   article: https://cromanpa94.github.io/phruta/articles
