+++ R/tree.dating.R
@@ -26,7 +26,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' sq.retrieve(
+#' sq.retrieve.direct(
 #'   clades = c("Felis", "Vulpes", "Phoca"),
 #'   species = "Manis_pentadactyla",
 #'   genes = c("ADORA3", "CYTB")
