+++ R/tree.roguetaxa.R
@@ -15,7 +15,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' sq.retrieve(
+#' sq.retrieve.direct(
 #'   clades = c("Felis", "Vulpes", "Phoca"),
 #'   species = "Manis_pentadactyla",
 #'   genes = c("ADORA3", "CYTB")
