+++ DESCRIPTION
@@ -46,7 +46,8 @@ Imports:
     msa,
     odseq,
     XML,
-    Rogue
+    Rogue,
+    doMC
 RoxygenNote: 7.2.0
 Suggests:
     testthat (>= 3.0.0),
