+++ NAMESPACE
@@ -18,6 +18,7 @@ export(tree.roguetaxa)
 import(DECIPHER)
 import(XML)
 import(ape)
+import(doMC)
 import(doParallel)
 import(doSNOW)
 import(foreach)
