+++ DESCRIPTION
@@ -46,8 +46,7 @@ Imports:
     msa,
     odseq,
     XML,
-    Rogue,
-    doMC
+    Rogue
 RoxygenNote: 7.2.0
 Suggests:
     testthat (>= 3.0.0),
@@ -56,7 +55,8 @@ Suggests:
     markdown,
     rmarkdown,
     spelling,
-    utf8
+    utf8,
+    doMC
 VignetteBuilder: knitr
 License: MIT + file LICENSE
 Config/testthat/edition: 3
