+++ R/acc.table.retrieve.R
@@ -20,7 +20,6 @@
 #' @import foreach
 #' @import doParallel
 #' @import doSNOW
-#' @import doMC
 #'
 #' @examples
 #' \dontrun{
