### 2022-05-31 0e33640
fixing the structure of gene.sampling.retrieve
