+++ R/sq.curate.R
@@ -27,6 +27,8 @@
 #'                Two possible options: "animals" or "plants."
 #' @param folder The name of the folder where the original sequences are
 #'               located (character).
+#' @param sqs.object A list of sequences generated from \code{sq.retrieve.indirect}. Only use if you're
+#'                   not interested in download sequences locally.
 #' @param removeOutliers Whether  \code{odseq:odseq} should be used to remove outliers
 #' @param minSeqs minimum number of sequences per locus
 #' @param threshold Relative to \code{odseq::odseq}. Only important if
@@ -67,6 +69,7 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
                       database = "gbif",
                       kingdom = NULL,
                       folder = "0.Sequences",
+                      sqs.object = NULL,
                       removeOutliers = TRUE,
                       minSeqs = 5,
                       threshold = 0.05,
@@ -87,35 +90,41 @@ sq.curate <- function(filterTaxonomicCriteria = NULL,
   if (is.null(folder)) stop("Folder where curated
                             sequences are saved must be provided")
 
-  if (!is.null(mergeGeneFiles)) {
 
-    mergedSeqs <- lapply(mergeGeneFiles, function(x){
-      refF <- list.files(folder, pattern = '.fasta')
-      targetF <- paste0(unlist(x), '.fasta')
-      if (all(targetF %in% refF)) {
+  if(is.null(sqs.object)){
 
-      seqs <- lapply(x, function(z) read.FASTA(paste0(folder,"/", z, '.fasta')))
 
-      #unlink(paste0(folder,"/", x, '.fasta')) #remove original files
-      file.rename(paste0(folder,"/", x, '.fasta'), paste0(folder,"/", x, '.original.non.merged'))
+    if (!is.null(mergeGeneFiles)) {
 
-      comS <- do.call(c,seqs)
-      comS[!duplicated(sub(".*? ", "", names(comS)))]
-      }else{
-        message("\nFiles ", paste(targetF, collapse = " AND "), ", expected to be merged, not found in ", folder,"\n")
-      }
-    })
+      mergedSeqs <- lapply(mergeGeneFiles, function(x){
+        refF <- list.files(folder, pattern = '.fasta')
+        targetF <- paste0(unlist(x), '.fasta')
+        if (all(targetF %in% refF)) {
 
-    names(mergedSeqs) <- names(mergeGeneFiles)
-    invisible(
-    lapply(seq_along(mergedSeqs), function(y){
-      if(!is.null(mergedSeqs[[y]])){
-    write.FASTA(mergedSeqs[[y]],
-                paste0(folder,"/", names(mergedSeqs)[y], ".fasta"))
-      }
-    })
-    )
-  }
+          seqs <- lapply(x, function(z) read.FASTA(paste0(folder,"/", z, '.fasta')))
+
+          #unlink(paste0(folder,"/", x, '.fasta')) #remove original files
+          file.rename(paste0(folder,"/", x, '.fasta'), paste0(folder,"/", x, '.original.non.merged'))
+
+          comS <- do.call(c,seqs)
+          comS[!duplicated(sub(".*? ", "", names(comS)))]
+        }else{
+          message("\nFiles ", paste(targetF, collapse = " AND "), ", expected to be merged, not found in ", folder,"\n")
+        }
+      })
+
+      names(mergedSeqs) <- names(mergeGeneFiles)
[…]
