+++ .github/workflows/pkgdown.yaml
@@ -23,7 +23,7 @@ jobs:
 
       - name: Query dependencies
         run: |
-          install.packages('remotes', 'devtools')
+          install.packages('remotes')
           saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
           writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
         shell: Rscript {0}
@@ -38,7 +38,7 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          install.packages("pkgdown", type = "binary")
+          install.packages("pkgdown", "devtools", type = "binary")
         shell: Rscript {0}
 
       - name: Install package
