+++ .github/workflows/pkgdown.yaml
@@ -38,7 +38,8 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_deps(dependencies = TRUE)
-          install.packages("pkgdown", "devtools", type = "binary")
+          install.packages("devtools")
+          install.packages("pkgdown", type = "binary")
         shell: Rscript {0}
 
       - name: Install package
