### 2022-06-27 ad60acc
Removing fig using vignette
