+++ tests/testthat.R
@@ -3,4 +3,5 @@ library(phruta)
 library(DECIPHER)
 library(Biostrings)
 library(msa)
+assign(".testMode", TRUE, envir = phruta:::pkg.env)
 test_check("phruta")
