+++ R/sq.partitionfinderv1.R
@@ -46,6 +46,16 @@ sq.partitionfinderv1 <- function(folderAlignments = "2.Alignments",
                                  FilePatterns = "Masked",
                                  folderPartitionFinder ="2.1.PartitionFinderv1",
                                  models = "all", run = TRUE) {
+
+
+  ##Over-writing?
+  if( !isTRUE(pkg.env$.testMode) ) {
+    UI <- readline(paste0("This function might overwrite ",
+                          "any PF folder", ". Are you sure you want to continue? (y/n)  "))
+    if(UI != 'y') stop('Exiting since you did not press y')
+  }
+
+
   files_fullNames <- list.files(folderAlignments, FilePatterns,
                                 full.names = TRUE)
   files <- list.files(folderAlignments, FilePatterns)
