+++ tests/testthat/helper-testMode.R
@@ -1,0 +1,1 @@
+assign(".testMode", TRUE, envir = phruta:::pkg.env)

[…]
