+++ vignettes/Exporting_data_phruta.Rmd
@@ -25,18 +25,19 @@ library(knitr)
 # Table of Contents
 
 1.  [To export or not export from `phruta`](#intro)
-3.  [From taxonomic names to sequence alignments, exporting](#paragraph1)
+2.  [From taxonomic names to sequence alignments, exporting](#paragraph1)
 
 ## To export or not export from `phruta` <a name="intro"></a>
 
-Note that assembling a molecular dataset with `phruta` might be performed entirely by saving objects to the workspace. However, sometimes it might be desirable to save the partial outputs to a particular folder. This tutorial will cover that specific situation. We will review how `phruta` can be used to save different `.fasta` and `.csv` files that are generated by different functions in `phruta`.
+Assembling a molecular dataset for particular target taxa with `phruta` can be performed almost entirely by saving objects into your workspace. This topic was covered in the introductory vignette to `phruta` (Using the `phruta` R package). However, in some situations, it might be desirable to save the outputs of different `phruta` functions to particular folders. This tutorial will cover that specific situation. Specifically, we will be reviewing how `phruta` can be used to export `.fasta` and `.csv` files that are generated in different steps of the pipeline.
 
+## From taxonomic names to sequence alignments, exporting data <a name="paragraph1"></a>
 
-## From taxonomic names to sequence alignments, exporting <a name="paragraph1"></a>
+In the introductory vignette to `phruta` (Using the `phruta` R package), we assembled a basic molecular dataset for inferring the phylogeny among three mammal genera. Let's recreate the same tutorial, but in this case, exporting the intermediate files that are created after using several of the functions. Note that the structure of this tutorial closely follows that of "Using the `phruta` R package". You should be able to follow this tutorial even without having reviewed the introductory vignette.
 
-Let's create a phylogeny for a few mammal clades. Assume that we need to build a tree for the following three genera: *Felis, Vulpes*, and *Phoca*. All three genera are within Carnivora. Both *Felis* and *Vulpes* are classified in different superfamilies within Fissipedia and *Phoca* is part of another suborder, Pinnipedia. We're going to root our tree with another mammal species, a Chinese Pangolin ( *Manis pentadactyla*). Users can select additional target species, however, for simplicity, we will run the analyses using three genera in the ingroup and a single outgroup species.
+Please assume that we are interested in building a phylogenetic tree for the following three genera: *Felis, Vulpes*, and *Phoca*. All these three genera are classified within the Carnivora, a mammalian order. Both *Felis* and *Vulpes* are classified in different superfamilies within the Fissipedia. Finally, *Phoca* is part of another suborder, Pinnipedia. We're going to root our tree with another mammal species, a Chinese Pangolin (*Manis pentadactyla*). Users can select additional target species and clades. However, for simplicity, we will run the analyses using three genera in the ingroup and a single outgroup species.
 
-So far, we have decided the taxonomic make of our analyses. However, we also need to determine the characteristics of the molecular dataset that we would like to assemble. Fortunately, mammals are extensively studied and a comprehensive list of potential gene regions to be analized is already available. For instance, we could use same gene regions as in [Upham et al (2009)](https://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.3000494#sec030). Nevertheless, for this tutorial, we will simply try to find the gene regions are well sampled for the target taxa. I believe that figuring out the best sampled gene regions in genbank, instead of providing names, is potentially more valuable when working with poorly studied groups. Before we move on, please make sure that you you have [set a working directory for this project](http://rfunction.com/archives/1001) as all the files will be saved to the current working directory.
[…]
