+++ vignettes/Using_phruta.Rmd
@@ -22,7 +22,6 @@ start_time <- Sys.time()
 library(knitr)
 ```
 
-
 # Table of Contents
 
 1.  [Clade names to sequence alignments](#paragraph1)
@@ -40,11 +39,9 @@ Let's start by loading `phruta`!
 library(phruta)
 ```
 
-
 ```{r setup, echo=FALSE, include = FALSE}
 assign(".testMode", TRUE, envir = phruta:::pkg.env)
 ```
-
 
 Now, let's look for the gene regions that are sampled for our target taxa. For this, we will use the `gene.sampling.retrieve` function in `phruta`. The resulting `data.frame`, named `gs.seqs` in this example, will contain the list of full names for genes sampled in genbank for the target taxa.
 
@@ -59,9 +56,9 @@ For the search terms, `phruta` was able to retrieve the names for `r nrow(gs.seq
 kable(head(gs.seqs), caption = "Gene sampling for the search including the genera Felis, Vulpes, Phoca as the ingroup taxa. Manis pentadactyla was used as outgroup.")
 ```
 
-The `gene.sampling.retrieve` provides an estimate of the number of species in genbank (matching the taxonomic criteria of the search term) that have sequences for a given gene region. However, this estimate is only as good as the annotations for genes deposited in genbank. 
+The `gene.sampling.retrieve` provides an estimate of the number of species in genbank (matching the taxonomic criteria of the search term) that have sequences for a given gene region. However, this estimate is only as good as the annotations for genes deposited in genbank.
 
-From here, we will generate a preliminary table summarizing accession numbers for the combination of taxa and gene regions that we're interested in sampling. In fact, not all these accession numbers are expected to be in the final (curated) molecular dataset. Using the `acc.table.retrieve` function, we will assemble a species-level summary of accession numbers (hence the `speciesLevel = TRUE` argument). For simplicity, this tutorial will focus on sampling gene regions that are sampled in >30% of the species (`targetGenes` data.frame).
+From here, we will generate a preliminary table summarizing accession numbers for the combination of taxa and gene regions that we're interested in sampling. In fact, not all these accession numbers are expected to be in the final (curated) molecular dataset. Using the `acc.table.retrieve` function, we will assemble a species-level summary of accession numbers (hence the `speciesLevel = TRUE` argument). For simplicity, this tutorial will focus on sampling gene regions that are sampled in \>30% of the species (`targetGenes` data.frame).
 
 ```{r  results = 'hide'}
 targetGenes <- gs.seqs[gs.seqs$PercentOfSampledSpecies > 30,]
@@ -105,8 +102,7 @@ Now, we still need to retrieve all the sequences from the accessions table gener
 sqs.downloaded <- sq.retrieve.indirect(acc.table = acc.table, download.sqs = FALSE)
 ```
 
-
[…]
