+++ vignettes/phruta_targetgenes.Rmd
@@ -22,23 +22,21 @@ start_time <- Sys.time()
 library(knitr)
 ```
 
+In this vignette, we will be retrieving sequences from genbank using `phruta`. We will need a list of taxa and the names of particular gene regions that need to be sampled. If you have no idea on what genes you should sample - `phruta`'ve got your back. Just check the "Introduction to the \`phruta\` R package" and "To export or not export \`phruta\` outputs" vignettes.
 
-In this vignette, we will be retrieving sequences directly from genbank using a list of taxa and particular gene regions. Given that we already have some target genes in mind, we can skip `gene.sampling.retrieve`. We won't generate an accession table using `acc.table.retrieve` at the beginning, mostly because the sampling that we're using in `phruta` this time, based on `sq.retrieve.direct`, is less flexible than the one outlined before for `sq.retrieve.indirect`. As noted before, `sq.retrieve.direct` does its best to directly (i.e. without potential input from the user) retrieve sequences for a target set of taxa and set of gene regions. You should be able to catch errors using `sq.retrieve.indirect` but mistakes will be harder to spot and fix if you're using `sq.retrieve.direct`. 
+Now, given that we already have some target genes in mind, we can avoid using `gene.sampling.retrieve`. We won't generate an accession table using `acc.table.retrieve` at the beginning, mostly because the sampling that we're using in `phruta` this time. Therefore, `sq.retrieve.direct` is less flexible than `sq.retrieve.indirect`. As noted before, `sq.retrieve.direct` does its best to directly (i.e. without input from the user) retrieve sequences for a target set of taxa and set of gene regions. You are more likely to catch errors using `sq.retrieve.indirect`. However, mistakes will be harder to spot and fix when using `sq.retrieve.direct`.
 
-Let's start by loading `phruta`!
+Let's first start by loading `phruta`!
 
 ```{r eval = TRUE, warning = TRUE, message = TRUE, progress = TRUE}
 library(phruta)
 ```
 
-
 ```{r setup, echo=FALSE, include = FALSE}
 assign(".testMode", TRUE, envir = phruta:::pkg.env)
 ```
 
-
-Now, let's download gene sequences for taxa in _Felis_, _Vulpes_, and _Phoca_. We will be sampling only two gene regions using `sq.retrieve.direct`. Note that we're going to be using the list of species for the target genera from the gbif taxonomic backbone (hence `db = "gbif"`).
-
+Let's download gene sequences for the taxa in *Felis*, *Vulpes*, and *Phoca*. Tjhis species-level sampling corresponds with the one in other vignettes. We will be sampling only two gene regions using `sq.retrieve.direct`. Note that we're going to be using the list of species for the target genera from the gbif taxonomic backbone (hence `db = "gbif"`). There are more databases available for this (please check `taxize::downstream`).
 
 ```{r results = 'hide'}
 sq.retrieve.direct(
@@ -49,9 +47,9 @@ sq.retrieve.direct(
 )
 ```
 
-This line of code will create a folder `0.Sequences` in your working directory. Note that `sq.retrieve.direct` does not return objects to the environment. The rest of this tutorial follow the same structure as the `Using the `phruta` R package` vignette. 
+The code above will create a folder `0.Sequences` in your working directory. Note that `sq.retrieve.direct` does not return objects to the environment. The rest of this tutorial follow the same structure as the "To export or not export \`phruta\` outputs" vignette.
 
[…]
