### 2022-06-29 5045fe3
update other functions pkgdown
