+++ _pkgdown.yml
@@ -40,6 +40,12 @@ reference:
   contents:
   - starts_with("tree")
   - sq.partitionfinderv1
+- title: "Other functions"
+  contents:
+  - SW.phruta
+  - acc.retrieve
+  - taxonomy.retrieve
+
 
 articles:
 - title: Tutorials (English)
