### 2022-06-29 b616bae
Fixing internal functions
