+++ man/SW.phruta.Rd
@@ -23,4 +23,4 @@ Scholl, Joshua P., and John J. Wiens.
 across the Tree of Life." Proceedings of the Royal
 Society B: Biological Sciences 283.1838 (2016): 20161334.
 }
-\keyword{datasets}
+\keyword{internal}
