+++ man/taxonomy.retrieve.Rd
@@ -48,3 +48,4 @@ taxonomy.retrieve(
 )
 }
 }
+\keyword{internal}
