+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(acc.retrieve)
 export(acc.table.retrieve)
 export(gene.sampling.retrieve)
 export(getListConstraints)
@@ -10,7 +9,6 @@ export(sq.curate)
 export(sq.partitionfinderv1)
 export(sq.retrieve.direct)
 export(sq.retrieve.indirect)
-export(taxonomy.retrieve)
 export(tree.constraint)
 export(tree.dating)
 export(tree.raxml)
