+++ vignettes/Exporting_data_phruta.Rmd
@@ -140,7 +140,7 @@ library(ape)
 files <- list.files("2.Alignments", pattern = "Raw_",full.names = TRUE)
 files.main <- list.files("2.Alignments", pattern = "Raw_",full.names = FALSE)
 sampleALNs <- lapply(files, read.FASTA); names(sampleALNs) <- files
-jpeg("EDF.Raw.jpeg", width = 8, height = 5, units = "in", res = 300)
+jpeg("EDF.Raw.jpeg", width = 8, height = 5, units = "in", res = 72)
 par(mfrow = c(ceiling(length(sampleALNs)/2),2))
 for (i in seq_along(sampleALNs)) {
   image(sampleALNs[[i]], main = files.main[i], show.labels = FALSE, xlab = "Position",
@@ -158,7 +158,7 @@ Now, the masked alignments...!!
 files <- list.files("2.Alignments", pattern = "Masked_",full.names = TRUE)
 files.main <- list.files("2.Alignments", pattern = "Masked_",full.names = FALSE)
 sampleALNs <- lapply(files, read.FASTA); names(sampleALNs) <- files
-jpeg("EDF.Cur.jpeg", width = 8, height = 5, units = "in", res = 300)
+jpeg("EDF.Cur.jpeg", width = 8, height = 5, units = "in", res = 72)
 par(mfrow = c(ceiling(length(sampleALNs)/2),2))
 for (i in seq_along(sampleALNs)) {
   image(sampleALNs[[i]], main = files.main[i], show.labels = FALSE, xlab = "Position",
@@ -181,4 +181,5 @@ In total, this vignette took `r round(end_time-start_time)` minutes to render in
 ```{r include = FALSE}
 #Deleting the extra elements...
 unlink(c("0.Sequences", "1.CuratedSequences", "2.Alignments"), recursive = TRUE)
+#unlink(c("EDF.Cur.jpeg", "EDF.Raw.jpeg"), recursive = TRUE)
 ```
