### 2022-06-30 #458 @cromanpa94
Hi friends,

Sorry for the delay...(I think) I ended up making too many improvements to the package + had to solve several new issues for passing the `R-CMD-check`s in `GitHub`...below are my responses to your comments (also roughly summarized in [here](https://github.com/cromanpa94/phruta/compare/48e8f6b...c2f8ceb)). Thank you @annakrystalli, @fboehm, and @raynamharris for your feedback!!

### Comments from `annakrystalli`

- [x] .DS_Store still in the R directory. Just noticed that .Rhistory and .Rproj.user are being version controlled. These will need removing from git also. As you may want to keep the files locally (just don't want them tracked by git), here's a useful post on how to do that. It's best to add .Rhistory and .Rproj.user/* to your .gitignore afterwards.

****Thank you so much for this comment!! I'm now following these recommendations.** See https://github.com/cromanpa94/phruta/commit/30bffcac78eb1f20f3777aa7189c7e6ac750c873 and https://github.com/cromanpa94/phruta/commit/f3a468f93fcd92e71e4d69eb0ade561738a9523d.

#### notes for reviewers

- [ ] One note on setwd. I appreciate that in some cases it might be necessary and that restoring it with on.exit() is a good way. But I do wonder whether it is necessary and couldn't instead be handled by having an explicit parent directory argument set to default to "." (equivalent to the current working directory) and better use of the file argument in ips::raxml(). The important thing here is that it is clear to the user where files will be written out and ideally give them the option to override the default.

****Thanks for pointing this out. I have spent a significant amount of time figuring out a solution for this aspect. I couldn't find an optimal (and efficient) alternative to what I'm doing right now. However, I'm still doing some testing locally to check if moving the files that are created using `ips::raxml` from the current working directory into `3.phylogeny` is an optimal solution vs changing the working directory.**

- [x] Another thing to think about is the deletion prior to re-writing of complete folders. You will likely have trouble submitting to CRAN with the current approach because it deletes existing files without asking the user for confirmation. This goes against CRAN policy, in particular:

> Packages should not write in the user’s home filespace (including clipboards), nor anywhere else on the file system apart from the R session’s temporary directory (or during installation in the location pointed to by TMPDIR: and such usage should be cleaned up). Installing into the system’s R installation (e.g., scripts to its bin directory) is not allowed. Limited exceptions may be allowed in interactive sessions if the package obtains confirmation from the user.

You might want to use/explore how `usethis::write_over()` function handles this and include an overwrite argument that y default is set to FALSE but can be set to overwrite folders explicitly by the user.

****I changed the structure of several of the basic functions in `phrura` to allow users to save the output of certain functions to the workspace. These changes involve the the two `sq.retrieve.*` functions, along with `sq.aln` and `sq.curate`. I also created a new vignette explaining the export functionality (`Exporting_data_phruta.RMD`), which is not the default behavior for the four main functions in `phruta`. If the user decides to write files into their working directory, the latest version of `phruta` also generates prompts that confirm if the user is Ok with (1) writing files to their working direcotry and (2) potentially overwriting existing folders. Finally, I suspect `usethis::write_over()` does not support exporting files in the variety of formats that `phruta` deals with. **

- [x] In any case, I won't hold up the review process for these points and will start looking for reviewers. However, it gives you some time to consider them and I will point these out to the reviewers so we can get their thoughts too and resolve by the end of the process. 👍😊

****Thank you so much!! I did my best to solve these comments during this first round of reviews. However, please let me know if you have any other suggestions!**

### Reviews from `raynamharris`

#### Package Review

- [x] Installation: Installation succeeds as documented. Partially. The R package itself installed just fine. RAxML installed just fine, thanks to an easy conda install option. I wasn't able to download and successfully install PATHd-8 or treePL, so instead of running tree.dating(), I used ape's read.tree() function and plotted the tree with plot()

****I’m sorry to hear this 😞 I’m not 100% what happened with `PATHd-8` or `treePL` on your end but I'd be happy to help if necessary. The functions calling these `PATHd-8` or `treePL` are optional (i.e. for users that are able to install them correctly in their machine). The rest of the functions to retrieve, curate, and align are the main ones in `phruta`. However, I have included additional details in the `README` on how to install the software that is external to `phruta`.**

****Excellent! Anna, Rayna, and Frederick were added as reviewers. See https://github.com/cromanpa94/phruta/commit/a74dc95972a45a46ffafd63b0b18c10f158e1bb2.**

Review Comments
The author wrote the phruta R package to simplify the basic phylogenetic pipeline. As promised, this R packages brings the functionality of multiple phylogenetic tools into the same environment where they can be used in a pipeline. The files are clearly organized and have standard formats, which makes it easy to use in combination with other tools outside the fruta package. sq.retrieve is my favorite function from this package. I wish I had this 10 years ago when I was searching for sequences in a webbrowser one at a time. I can clearly see how I could incorporate this into a data discovery workflow. I also think would be a useful tool to use when teaching, to allow students to quickly make trees with their genes and species of interest.

****Thank you so much for saying this. I have actually created another function (`seq.retrieve.indirect`) that allows for a better control over the sequences that are being retrieved (relative to `sq.retrieve.direct`, equivalent to the older `sq.retrieve` mentioned in Rayna's comment). The difference between both `seq.retrieve.indirect` and `seq.retrieve.direct` is now covered in two new vignettes:**

1. **`phruta` will find the "best-sampled" genes given your target taxa:** https://cromanpa94.github.io/phruta/articles/Using_phruta.html
2. **`phruta` will download for a combination of target genes and taxa:**  [Add]

**Note that if users don't know what genes are widely sampled for a given set of taxa, `phruta` can also make suggestions on this aspect through the `gene.sampling.retrieve`. I really hope that the option of retrieving both species and genes that are of potential interest for a given study makes `phruta` especially handy for many folks.**

- [x] The tutorial was easy to follow. However, I found it a little difficult to find detailed information about the functions. It would be nice if some of the descriptions for the functions were listed or described on the home page of the website, so the uses knows more about the capability of the package.

****I have included additional details on the functions both in the `README` and directly in the website. See https://github.com/cromanpa94/phruta/commit/f3555851ebdf96f2dd65b42eb70cc18aa97629d9. For instance, I provide additional details on how the `pipeline` works in the `What is phruta?` [vignette](https://cromanpa94.github.io/phruta/articles/what_is_phruta.html). There's also a `Reference` section in the [website](https://cromanpa94.github.io/phruta/reference/index.html) that defines each of the functions in `phruta`.**

- [x] A statement of need: clearly stating problems the software is designed to solve and its target audience in README.

****Thanks for making this comment. I have included a new statement in the `README` indicating the target audience for `phruta`** See https://github.com/cromanpa94/phruta/commit/949db90c3f477a572a07ff8beeb3c49e07140cbf

- [x] Installation instructions: for the development version of package and any non-standard dependencies in README

****This is a great point. I will add this information once I create a branch with the development of `phruta`.** I did this at some point but decided to tempoerally focus on the main branch given the major changes in this review: https://github.com/cromanpa94/phruta/commit/a41eaac2d1a9c3b3dc0f86cca71951f89ef3fcba.

 - [x] Vignette(s): demonstrating major functionality that runs successfully locally

****The first version of `phruta` included only a single major vignette. The latest release includes a total of seven different vignettes explaining alternative aspects of `phruta`. When external software is not requiered (e.g. `ips::raxml` is not used), all blocks of code are evaluated. Note tha each vignette takes >10 min to render...***

- [x] Function Documentation: for all exported functions. Examples: (that run successfully locally) for all exported functions

****Thank you!! All the exported functions have associated examples.**

- [x] Performance: Any performance claims of the software been confirmed.

****I didn’t actually compare performance to anything else (because there’s nothing similar). However, I did include estimates of times to render the main vignettes of the package. I also describe alternative sorfware to `phruta` in the `README`.**

### Reviews from `fboehm`

#### Package Review

- [x] A statement of need: clearly stating problems the software is designed to solve and its target audience in README. I think a sentence - or part of a sentence - could be added to clarify the intended audience. The problem that the software addresses is clearly stated.

****Excellent – I added this information to the `README`**. See https://github.com/cromanpa94/phruta/commit/949db90c3f477a572a07ff8beeb3c49e07140cbf.

- [x] Installation instructions: for the development version of package and any non-standard dependencies in README
You might need to move or copy the installation instructions for RAxML, PATHd-8, and treePL to the README.md.

****Great suggestion! This change was added to the latest version of the `README`**. See https://github.com/cromanpa94/phruta/commit/59bdea9fd946aaa7851e04ac8dfb88f9e1bfdfb6. Note that, as indicated above, I will create a new development branch once we have agreed on the main changes. I will add the relevant instructions to the `README` as soon as I create the new branch!

- [x] Vignette(s): demonstrating major functionality that runs successfully locally. While the vignettes can be rendered successfully with rmarkdown, it seems that many or all of the code chunks are not evaluated when rendering code. Is this intentional? Why? If computing time is a concern, it might be better to read results into the vignette code. For example, add a note to the vignette that the computing time is too long to actually run the code, so you're going to read the saved results - that you could add to the package itself - into the R session. Be sure to continue to include the code that's already in the vignette.

****Thanks for pointing this out. The main vignettes that do not requiere external software (e.g. through `ips::raxml`) are not being evaluated in full, without relevant chunks set to `eval = FALSE`.  Note that building the most important vignettes take >10 min. I really hope this is not an issue for users interested in rendering the vignettes locally.**

- [x] Function Documentation: for all exported functions. It looks like all exported functions have sufficient Roxygen documentation.

****Thanks!**

- [x] Examples: (that run successfully locally) for all exported functions. It looks like the examples all are bracketed with code to designate that they not be run. Why is this?

****I’m preventing examples from running given that the computation time is, most of the time, above 5 min. There’s general information on the time that takes most of the vignettes to render in the [website](https://cromanpa94.github.io/phruta/articles/Using_phruta.html).**

- [x] Community guidelines: including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

****Already part of the initial submission!**

- [x] Installation: Installation succeeds as documented. The installation of the R package itself succeeds as documented. However, see my comments above about the installation of RAxML, PATHd-8, and treePL.

****Correct. This was added to the README.**

- [x] Functionality: Any functional claims of the software been confirmed. I'm not aware of any functional claims. Did I overlook something?

****Thanks for pointing this out. There's now a new section in the `README` that focuses on comparing `phruta` to similar packages in `R` and `python`.**

- [x] Performance: Any performance claims of the software been confirmed. I'm also not aware of performance claims.

****I provide estimates of the time that it takes to render the vignettes. However, the performance of `phruta` necessarily depends on the searchers that the user is trying to conduct.**

- [x] Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine. What is the purpose of the "tests/testthat/_snaps" directory? It contains a single md file. The "tests/testthat" subdirectory contains seven subdirectories that each contain data files. Is this the best place to store data files for tests? Or would it be more sensible to store the files used in tests with the other data files in this package? I haven't seen other packages that store data in the "testthat" directory.

****I removed all the partial results that were created when running `testthat`, including`tests/testthat/_snaps`. Thanks for pointing this out!**

- [x] I see also that codecov is at 75%, according to the badge on the Github repository. The collection of tests feel incomplete. For example, at line 14 of file "tests/testthat/testthat.R", the first argument is "Error in clades and species". Why is an error expected here? And which error is it? Do you want to specify the expected error message? At line 57, I don't know what "Max sequence as character" means. Can you elaborate this first argument?

****I have included more details to the relevant files in testthat, including the ones listed by the reviewer.**

- [x] Also, I might add that the organization of the tests, where all tests are in a single file, made it a little harder to assess the coverage of tests. Would it be better to organize the tests with a single file per exported function?

**** I have followed this suggestion from the reviewer. All the test scripts are organized by a given task (e.g. running the full pipeline) or function (e.g. run `sq.curate`).**

- [x] What happens if a database query returns no results? Is this a problem, and is it detected with your code?

****This is a good point. The functions searching (e.g. `gene.sampling.retrieve`, `acc.table.retrieve`) or retriving sequences (e.g. `sq.retrieve.direct`) from genbank print to the console information on the searches. For instance, `acc.table.retrieve` prints to the console `Sequence found for gene 12S ribosomal RNA and organism Vulpes` when the code in `To export or not export `phruta` outputs` is run. This information is not shown in the vignettes but is printed to the user's console when `phruta` is run in interactive mode. Finally, note that if a database query doesn't return results, functions for retrieving data will simply not return or write anything.**

phruta is a really nice package that accesses several stand-alone software packages from R for use in phylogenetics analysis. I encourage the authors to elaborate on the suite of unit tests and to consider reorganizing the suite of tests from one file into multiple files, perhaps one file per exported function.

****Thank you so much for your comments and kind words!!**

- [x] doc directory - is it needed? I notice that there are two directories, doc and docs, with similar names. What is the motivation for having the two distinct directories with these names? Upon inspecting the contents, it looks like perhaps pkgdown generates the docs directory. This leaves me wondering about the need for the doc directory. doc seems to contain only vignettes, despite the presence of a vignettes directory. Is it possible to delete the doc directory and its contents?

****`doc` is removed from the latest submission! At some point, I wanted to provide the vignettes in `PDF`, `RMD`, and `R` formats but I ended up just liking the way vignettes were rendered in a website (`pkgdown`).**

- [x] Installing RAxML, PATHd-8, and treePL - is installation platform-specific? It looks like the authors assume a Mac operating system when writing the installation instructions in the file vignettes/Brief_intro_phruta.html. With much scientific computing occurring on machines with a linux operating system, would it be sensible to specify installation instructions for linux users? Are these software packages also available for Windows?

****This is an excellent point. The instructions on how to install `RAxML`, `PATHd-8`, and `treePL` are mostly targetted towards MacOS users. Linux users can generally follow similar instructions (e.g. `brew`) and this is now indicated in the `README`. Instructions on how to install `RAxML` in Windows and Linux are posted in the official website to the software (now linked to the `phruta`'s README). `treePL` can be installed using `brew` in windows, MacOS, and Linux. This information is also provided in the latest version of the `README`. Finally, I provide instructions on how to install `PATHd-8` in both Linux and MacOS, along with a link to the compiled version of the software for Windows.**

- [x] Confusing statement in vignette, vignettes/Brief_intro_phruta.html. The sentence "You may also want to check if RAxML is called using raxmlHPC or something else." is confusing. Can you provide additional context for this statement? Why might RAxML be called something else?

****Thank you so much. I added a few other details and referred the users to the argument in to `ips::raxml` for additional clarification.** See https://github.com/cromanpa94/phruta/commit/e36a916a086096deeb2c2b9f8a684274a627d919.** I also provide the same details in the `README` file to `phruta`.**

- [x] phruta is misspelled as pruta in vignettes/Brief_intro_phruta.html. Under section "Functions in phruta", the text for the third bullet uses pruta.

****Thanks for pointing this out. I fixed several of these issues in the vignettes**. See https://github.com/cromanpa94/phruta/commit/cf7874cc1ed69cbfbace36565ef3d050b7653ba4.

- [x] Delete remaining .DS_Store files from subdirectories. R/ contains such a file, for example.

****I deleted all the .DS_Store files!**

- [x] I'm less familiar with the use of global variables. Are these necessary and helpful here? Why?

****Great question. I’m defining global variables in `R/globals.R` to avoid the “undefined global variables” flag when the checks are run. In the latest version of the package, one global variable is also created when the package is loaded `.testMode`. This variable is used to avoid requiering any inputs from the user when the package is being tested or vignettes are being rendered.**

## Additional comments

[…]
