### 2022-12-26 b8a8d9b
merging the latest version of dev (#18)
* Updates based on ms

* n searches per batch

* Passing to acc.table.retrieve

* Conditioning partitioned analyses

* generate folder sq.retrieve.indirect

* Updating errors

* Updating the website and fixing seq retrieval
