+++ R/sq.retrieve.indirect.R
@@ -34,16 +34,20 @@
 
 sq.retrieve.indirect <- function(acc.table, download.sqs = FALSE){
 
+    if (download.sqs) {
+      unlink("0.Sequences", recursive = TRUE)
+      dir.create("0.Sequences")
+    }
 
-  unlink("0.Sequences", recursive = TRUE)
-  dir.create("0.Sequences")
 
  su <- pblapply(unique(acc.table$gene), function(x){
     acc.table.sub <- acc.table[acc.table$gene == x,]
     seqs <- read.GenBank(acc.table.sub$Acc, species.names = TRUE)
 
     names(seqs) <- paste0(names(seqs), " ", acc.table.sub$Species)
-    if(download.sqs){
+
+    
+    if (download.sqs) {
 
       ##Over-writing?
       if( !isTRUE(pkg.env$.testMode) ) {
