+++ README.md
@@ -7,13 +7,11 @@
 
 # The `phruta` `R` package <a href='https://cromanpa94.github.io/phruta'><img src='man/figures/logo.png' align="right" height="300" /></a>
 
-### Assembling phylogenetic trees from taxonomic names.
+### Assembling phylogenetic trees from taxonomic names
 
 ## What is `phruta`?
 
-The `phruta` R package is designed to simplify the basic phylogenetic pipeline. Specifically, all code is run within the same program and data from intermediate steps are saved in independent folders. Furthrmore, all code is run within the same environment which increases the reproducibility of your analysis. `phruta` retrieves gene sequences, combines newly downloaded and local gene sequences, and performs sequence alignments. 
-
-`phruta` is also able to perform basic phylogenetic inference under `RAxML` on the resulting sequence alignments. The current release allows users to conduct tree dating based on secondary calibrations. `phruta` is essentially a wrapper for alternative R packages and software.
+The `phruta` R package is designed to simplify the basic phylogenetic pipeline. All the code is run within the same program and data from intermediate steps are saved in independent folders (optional). `phruta` retrieves gene sequences, combines newly downloaded to local gene sequences, performs sequence alignments, and basic phylogenetic inference. 
 
 ## Who should consider using `phruta`
 
@@ -44,7 +42,16 @@ library(devtools)
 install_github("cromanpa94/phruta")
 ```
 
-Please make sure that the R packages `msa`, `DECIPHER`, `Biostrings`, and `odseq` are correctly installed. 
+Please make sure that the R packages `msa`, `DECIPHER`, `Biostrings`, and `odseq` are correctly installed. If you are interested in using the development version of `phruta`, please install it using the following code:
+
+```
+library(devtools)
+install_github("cromanpa94/phruta", ref = "dev")
+```
+
+## Running `phruta` from shiny
+
+I have constructed a shiny app that hosts `phruta` and enables users to run the basic functions in a less-code intensive environment. The app, `salphycon` is currently available in the following [GitHub repo](https://github.com/cromanpa94/salphycon). The shiny app will be live at some point in 2023.
 
 
 ## Installing RAxML <a name="paragraph1"></a>
@@ -102,9 +109,10 @@ open /Applications/RStudio.app
 ```
 
 ## Dedication
-My package is dedicated to my mom and every single Black woman. I still have lots of things to learn from you. You will always have all my admiration.
 
[…]
