+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   phruta: phruta.html
   phruta_targetgenes: phruta_targetgenes.html
   usando_phruta: usando_phruta.html
-last_built: 2022-12-26T17:01Z
+last_built: 2022-12-26T20:19Z
 urls:
   reference: https://cromanpa94.github.io/phruta/reference
   article: https://cromanpa94.github.io/phruta/articles
