### 2023-01-07 #458 @cromanpa94
Hi @ropensci/blog-editors! I'm interested in writing a brief intro for `phruta`. I've been working on the blog posts in this [repo](https://github.com/cromanpa94/roweb3/tree/phruta_post). Looking forward to your suggestions!
