### 2023-01-09 #458 @steffilazerte
Hi @cromanpa94,

I took a quick look, I think this will be a nice addition to the blog, thanks! One thing I would suggest at this point, is to consider adding more 'story' details. Perhaps start with who you are, and why you wanted to write this package. Some other things to consider might be: What problems did it solve? Was it tricky? Did you learn anything interesting along the way? (You don't have to answer all, or any, of these questions, this is just to spark ideas :grin:)

Generally speaking, we're trying to ensure that posts on the blog aren't simply tutorials. There's always a story behind packages, and it's nice to give readers a bit of insight into that world. Feel free to have fun with it!

Feel free to open a PR (include the [blog guide checklist](https://blogguide.ropensci.org/authorchecklistpeer.html)). I'll do a 'formal' review and we can figure out a date for publication.

Thanks again!
